% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova}
\alias{anova}
\title{Analysis of Variance}
\usage{
anova(df, keep_data_object = FALSE)
}
\arguments{
\item{df}{data frame}

\item{keep_data_object}{logical to keep objects from aov and TukeyHSD}
}
\description{
Analysis of Variance
}
\examples{
data(gw_data)

wells <- c("MW-1", "MW-2", "MW-3", "MW-4")

params <- c("Calcium, dissolved", "Sodium, dissolved", "Sulfate, total")

gw_data \%>\%
  filter(location_id \%in\% wells, param_name \%in\% params) \%>\%
  anova()

anova_object <- gw_data \%>\%
  filter(location_id \%in\% wells, param_name \%in\% params) \%>\%
  anova(., keep_data_object = TRUE)

plot(anova_object$anova[[1]], 1)

}
