% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_limits.R
\name{assign_limits}
\alias{assign_limits}
\title{Function to assign EPA primary and seconday Maximum Contaminant Levels}
\usage{
assign_limits(df, column = "param_name")
}
\arguments{
\item{df}{dataframe of groundwater data in tidy format}

\item{column}{column to search for constituents}
}
\description{
Function to assign EPA primary and seconday Maximum Contaminant Levels
}
