% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox_transform.R
\name{boxcox_transform}
\alias{boxcox_transform}
\title{Boxcox transformation}
\usage{
boxcox_transform(df, combine_locations = FALSE,
  objective_name = "PPCC")
}
\arguments{
\item{df}{data frame}

\item{combine_locations}{logical}

\item{objective_name}{character string indicating what objective to use.
The possible values are "PPCC" (probability plot correlation coefficient;
the default), "Shapiro-Wilk" (the Shapiro-Wilk goodness-of-fit statistic),
and "Log-Likelihood" (the log-likelihood function).}
}
\description{
Boxcox transformation
}
