% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{boxplot}
\alias{boxplot}
\title{Function to plot boxplots for groundwater data}
\usage{
boxplot(df, x = "LOCATION_ID", y = "ANALYSIS_RESULT",
  lt_measure = "LT_MEASURE", group_var = "PARAM_NAME",
  scale_y_trans = "identity", fill = NULL, limit1 = NULL, limit2 = NULL,
  pnt = 2, short_name = FALSE, coord_flip = FALSE, legend_title = NULL)
}
\arguments{
\item{df}{groundwater data in tidy format}

\item{x}{column to be used for the x axis, default is location}

\item{y}{column to be used for the y axis, default is analysis result}

\item{lt_measure}{column for >, or < identifier}

\item{scale_y_trans}{type of transformation to use for y scale. Default is
"identity".  Built-in transformations include "asn", "atanh", "boxcox", 
"exp", "identity", "log", "log10", "log1p", "log2", "logit", "probability", 
"probit", "reciprocal", "reverse" and "sqrt".}

\item{fill}{column used to fill the variable}

\item{limit1}{column to be used to draw horizontal line}

\item{limit2}{column to be used to draw a second horizontal line}

\item{short_name}{If TRUE the constituent name will be abbreviated}

\item{coord_flip}{If TRUE the axes are flipped}

\item{legend_title}{Legend title for fill variable}
}
\description{
Function to plot boxplots for groundwater data
}
