% Generated by roxygen2 (4.0.2): do not edit by hand
\name{boxplot_by_param}
\alias{boxplot_by_param}
\title{Function to plot boxplots by parameter}
\usage{
boxplot_by_param(df, ...)
}
\arguments{
\item{df}{groundwater data}

\item{name}{If name = "short" the analyte will be abbreviated}

\item{flip_coords}{If TRUE the axes are flipped}
}
\description{
Function to plot boxplots by parameter
}

