% Generated by roxygen2 (4.0.1): do not edit by hand
\name{boxplot_by_well_grid}
\alias{boxplot_by_well_grid}
\title{Function to plot boxplots by location id}
\usage{
boxplot_by_well_grid(df, name = NULL, coord_flip = FALSE)
}
\arguments{
\item{df}{groundwater data}

\item{name}{If name = "short" the analyte will be abbreviated}

\item{flip_coords}{If TRUE the axes are flipped}
}
\description{
Function to plot boxplots by location id
}

