% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bray_curtis.R
\name{bray_curtis}
\alias{bray_curtis}
\title{bray_curtis}
\usage{
bray_curtis(site1, site2, sorensen_dice = FALSE)
}
\arguments{
\item{site1}{A vector containing a set of observations (the observations at the first site).}

\item{site2}{A vector containing a set of observations (the observations at the second site).}

\item{sorensen_dice}{If TRUE, the Sørensen-Dice similarity index (one minus Bray-Curtis dissimilarity index) is returned (FALSE by default)}
}
\value{
A number [0, 1].
}
\description{
Calculate the Bray-Curtis dissimilarity index (or Sorensen-Dice similarity index) between two sites (with site compositions given as vectors).
}
\details{
No dependencies.
}
\examples{
area1 <- c('bubble-tip anemone', 'bubble-tip anemone', 'beaded sea anemone',
           'sebae anemone', 'magnificent sea anemone')
area2 <- c('bubble-tip anemone', 'bubble-tip anemone', 'bubble-tip anemone',
           'corkscrew tentacle sea anemone')
bray_curtis(area1, area2)

}
