% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_person_regex.R
\name{check_non_person_regex}
\alias{check_non_person_regex}
\title{check_non_person_regex}
\usage{
check_non_person_regex()
}
\value{
Output in the console describing the results of the check.
}
\description{
Checks efficacy of \code{non_person_regex} in flagging non-person names.
}
\details{
\code{check_non_person_regex()} checks whether the pattern contained in \code{non_person_regex} successfully tags all of the non-person names in the \code{test_names} dataset as non-persons while skipping individuals' "distraction" names.

Requires: \code{stringi}
}
\examples{
check_non_person_regex()

}
