\name{compare_MCL}
\alias{compare_MCL}
\title{Function to compare result to EPA's MCL}
\usage{
compare_MCL(df, format = "summary", type = "all")
}
\arguments{
  \item{df}{dataframe of groundwater data that has been
  unit checked}

  \item{format}{default summary}

  \item{type}{default is all}
}
\description{
Function to compare result to EPA's MCL
}

