% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{concatenate}
\alias{concatenate}
\title{Concatenate strings and ignore missing values}
\usage{
concatenate(...)
}
\arguments{
\item{...}{several string values}
}
\value{
Concatenated strings given as input with space delimiter. A missing
value (\code{NA}) is simply ignored. If all niputs are \code{NA}'s, \code{NA} is
returned
}
\description{
Concatenate strings and ignore missing values
}
\examples{
concatenate("Hello", NA, "world", NA)
concatenate(NA, NA)
}
