% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{conf_int}
\alias{conf_int}
\title{Function to calculate confidence intervals}
\usage{
conf_int(df, analysis_result = "analysis_result",
  distribution = "distribution", method = "mvue", ci_type = "two-sided",
  conf_level = 0.95)
}
\arguments{
\item{df}{df data frame of groundwater data in tidy format}

\item{analysis_result}{the analysis result column}

\item{distribution}{the distribution column}

\item{method}{character string specifying the method of estimation.
Possible values are "mvue" (minimum variance unbiased; the default),
and "mle/mme" (maximum likelihood/method of moments).}

\item{ci_type}{character string indicating what kind of confidence interval
to compute. The possible values are ci_type="two-sided" (the default),
ci_type="lower", and ci_type = "upper}

\item{conf_level}{a scalar between 0 and 1 indicating the confidence level
of the prediction interval. The default value is conf_level = 0.95}
}
\description{
Function to calculate confidence intervals
}
