% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_data.R
\name{connect_manages}
\alias{connect_manages}
\title{Connect to MANAGES database}
\usage{
connect_manages(manages_path)
}
\arguments{
\item{manages_path}{Path to MANAGES Site.mdb file}
}
\description{
This function connects to the MANAGES database for the path supplied.
R must be in 32-bit mode.
}
