% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_data.R
\name{connect_manages3}
\alias{connect_manages3}
\title{Connect to a MANAGES 3.x Site.mdb database}
\usage{
connect_manages3(manages3_path)
}
\arguments{
\item{manages3_path}{manages_path Path to MANAGES Site.mdb file}
}
\description{
This function connects to the MANAGES database for the path supplied.
R must be in 32-bit mode.
}
