% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_data.R
\name{connect_manages4}
\alias{connect_manages4}
\title{Function to connect to MANAGES 4.0 database}
\usage{
connect_manages4(driver = "SQL Server", server, database)
}
\arguments{
\item{server}{server name}

\item{database}{database name}
}
\description{
Function to connect to MANAGES 4.0 database
}
