\name{connect_manages_spatial}
\alias{connect_manages_spatial}
\title{function to connect to MANAGES database and read spatial data}
\usage{
connect_manages_spatial(manages_path)
}
\arguments{
  \item{manages_path}{path to MANAGES database}
}
\description{
function to connect to MANAGES database and read spatial
data
}

