% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{constituents}
\alias{constituents}
\title{Function to return a list of all the constituents}
\usage{
constituents(df, PARAM_NAME)
}
\arguments{
\item{df}{data frame of groundwater data in the format with column names}

\item{param_name}{the column for the constituents}
}
\description{
Function to return a list of all the constituents
}
