\name{convert_mgL_to_meqL}
\alias{convert_mgL_to_meqL}
\title{function to convert geochemical plot data into meq/L}
\usage{
convert_mgL_to_meqL(df, Mg = df$`Magnesium, dissolved`, Ca = df$`Calcium,
  dissolved`, Na = df$`Sodium, dissolved`, K = df$`Potassium, dissolved`,
  Cl = df$`Chloride, total`, SO4 = df$`Sulfate, total`,
  HCO3 = df$`Alkalinity, total (lab)`)
}
\description{
function to convert geochemical plot data into meq/L
}

