\name{convert_units}
\alias{convert_units}
\title{Function to convert units}
\usage{
convert_units(df, element, from, to)
}
\arguments{
  \item{df}{dataframe of groundwater data}

  \item{element}{element name to convert}

  \item{from}{original units}

  \item{to}{new units}
}
\description{
Function to convert units
}

