% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_database}
\alias{create_database}
\title{Function to create a new database for water quality data}
\usage{
create_database(con)
}
\arguments{
\item{con}{connection to Postgresql database}
}
\description{
Function to create a new database for water quality data
}
