% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_gwps}
\alias{default_gwps}
\title{U.S. EPA Maximum Contaminant Level and RSL Lookup Table}
\value{
mcl data frame specific to CCR Program
}
\description{
U.S. EPA Maximum Contaminant Level and RSL Lookup Table
}
