% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_dist.R
\name{display_dist}
\alias{display_dist}
\title{display_dist}
\usage{
display_dist(x, width = 10)
}
\arguments{
\item{x}{A numeric vector (double or integer).}

\item{width}{The number of blocks comprising the printed distribution.}
}
\value{
Console output.
}
\description{
Displays the approximate shape of a distribution in the console (or other i/o system) using Unicode Block Elements glyphs.
}
\details{
\code{display_dist()} is motivated by \code{precis::histospark()} (Hadley Wickham) and \code{rethinking::precis()} (Richard McElreath).
It displays a slightly more-granular distribution than those functions, and the peak(s) of the distribution is easily identified via the
use of a distinctive Block Elements glyph.
}
\examples{
# Normal
set.seed(99)
x <- rnorm(1000, 0, 1)
display_dist(x, 10)

# Poisson
set.seed(99)
x <- rpois(1000, lambda = 6)
display_dist(x, 10)

}
