% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{dist}
\alias{dist}
\title{Function to return distribution based on using p-value}
\usage{
dist(x)
}
\arguments{
\item{x}{analysis_result of groundwater data}
}
\description{
Function to return distribution based on using p-value
}

