% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{dixon_flag}
\alias{dixon_flag}
\title{Function to flag outliers from Grubb's test for outliers}
\usage{
dixon_flag(df, x = "analysis_result", ...)
}
\arguments{
\item{x}{column of analysis results}

\item{...}{other arguments passed to dixon test}
}
\description{
Function to flag outliers from Grubb's test for outliers
}
