% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{dixon_test}
\alias{dixon_test}
\title{Dixon test for outliers}
\usage{
dixon_test(df, x = "analysis_result", type = 0, opposite = FALSE,
  two_sided = TRUE, group_by_location = FALSE,
  keep_data_object = FALSE)
}
\arguments{
\item{df}{data frame}

\item{x}{column of analysis results}

\item{type}{an integer specyfying the variant of test to be performed.
Possible values are compliant with these given by
Dixon (1950): 10, 11, 12, 20, 21. If this value is set to zero, a variant
of the test is chosen according to sample size
(10 for 3-7, 11 for 8-10, 21 for 11-13, 22 for 14 and more).
The lowest or highest value is selected automatically, and can be reversed
used opposite parameter.}

\item{opposite}{logical indicating whether you want to check not the value
with largest difference from the mean, but opposite
(lowest, if most suspicious is highest etc.)}

\item{two_sided}{treat test as two-sided. Default is TRUE.}

\item{group_by_location}{TRUE/FALSE}

\item{keep_data_object}{TRUE/FALSE to return entire htest object}
}
\description{
Dixon test for outliers
}
