% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{.boxplot}
\alias{.boxplot}
\title{Helper function to plot boxplots for groundwater data}
\usage{
.boxplot(df, x = "location_id", y = "analysis_result", coef = 3,
  lt_measure = "lt_measure", scale_y_trans = "identity",
  show_points = TRUE, fill = NULL, limit1 = NULL, limit2 = NULL,
  pnt = 2, short_name = FALSE, coord_flip = FALSE,
  legend_title = NULL)
}
\description{
Helper function to plot boxplots for groundwater data
}
