% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_plot.R
\name{.get_corr_data}
\alias{.get_corr_data}
\title{Function to create a correlation matrix plot}
\usage{
.get_corr_data(df, location_id = "location_id",
  sample_date = "sample_date", param_name = "param_name",
  analysis_result = "analysis_result", constituents = NULL,
  sample_locations = NULL)
}
\description{
Function to create a correlation matrix plot
}
