% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper_plot.R
\name{.get_piper_ions}
\alias{.get_piper_ions}
\title{Function to gather major ions for piper plot.}
\usage{
.get_piper_ions(df, location_id = "location_id",
  sample_date = "sample_date", param_name = "param_name",
  analysis_result = "analysis_result", default_unit = "default_unit",
  x_cation = "Calcium, dissolved", y_cation = "Magnesium, dissolved",
  z_cation = c("Sodium, dissolved", "Potassium, dissolved"),
  x_anion = c("Chloride, total", "Fluoride, total"),
  y_anion = "Alkalinity, total (lab)", z_anion = "Sulfate, total",
  total_dissolved_solids = NULL)
}
\description{
Function to gather major ions for piper plot.
}
