% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiff_plot.R
\name{.get_stiff_ions}
\alias{.get_stiff_ions}
\title{Function to gather major ions for stiff diagram.}
\usage{
.get_stiff_ions(df, location_id = "location_id",
  sample_date = "sample_date", param_name = "param_name",
  analysis_result = "analysis_result", default_unit = "default_unit",
  calcium = "Calcium, dissolved", magnesium = "Magnesium, dissolved",
  sodium = "Sodium, dissolved", potassium = "Potassium, dissolved",
  chloride = "Chloride, total", alkalinity = "Alkalinity, total (lab)",
  sulfate = "Sulfate, total", total_dissolved_solids = NULL)
}
\description{
Function to gather major ions for stiff diagram.
}
