% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{.sp_plot}
\alias{.sp_plot}
\title{Helper function for sp_plot}
\usage{
.sp_plot(df, x = "analysis_result", color = "location_id",
  sample_interval = "semi-annually", group_var = NULL,
  facet_var = NULL)
}
\description{
Helper function for sp_plot
}
