% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiff_plot.R
\name{.stiff_plot}
\alias{.stiff_plot}
\title{Helper function for Stiff Diagrams.}
\usage{
.stiff_plot(df, location_id = "location_id",
  sample_date = "sample_date", param_name = "param_name",
  analysis_result = "analysis_result", default_unit = "default_unit",
  magnesium = "Magnesium, dissolved", calcium = "Calcium, dissolved",
  sodium = "Sodium, dissolved", potassium = "Potassium, dissolved",
  chloride = "Chloride, total", sulfate = "Sulfate, total",
  alkalinity = "Alkalinity, total (lab)",
  total_dissolved_solids = NULL, facet_var = "sample_date",
  lines = FALSE, cex = 1)
}
\description{
Helper function for Stiff Diagrams.
}
