% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper_plot.R
\name{.transform_piper_data}
\alias{.transform_piper_data}
\title{Function to transform data for piper plot}
\usage{
.transform_piper_data(df, location_id = "location_id",
  sample_date = "sample_date", x_cation = "Calcium, dissolved",
  y_cation = "Magnesium, dissolved", z_cation = c("Sodium, dissolved",
  "Potassium, dissolved"), x_anion = c("Chloride, total",
  "Fluoride, total"), y_anion = "Alkalinity, total (lab)",
  z_anion = "Sulfate, total", total_dissolved_solids = NULL)
}
\description{
Function to transform data for piper plot
}
