% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoeller_plot.R
\name{.transform_schoeller_data}
\alias{.transform_schoeller_data}
\title{Help function to transform data to schoeller coordinates}
\usage{
.transform_schoeller_data(df, location_id = "location_id",
  sample_date = "sample_date", param_name = "param_name",
  analysis_result = "analysis_result", default_unit = "default_unit",
  calcium = "Calcium, dissolved", magnesium = "Magnesium, dissolved",
  sodium = "Sodium, dissolved", potassium = "Potassium, dissolved",
  chloride = "Chloride, total", alkalinity = "Alkalinity, total (lab)",
  sulfate = "Sulfate, total")
}
\description{
Help function to transform data to schoeller coordinates
}
