% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiff_plot.R
\name{.transform_stiff_data}
\alias{.transform_stiff_data}
\title{Function to transform data to Stiff diagram coordinates}
\usage{
.transform_stiff_data(df, location_id = "location_id",
  sample_date = "sample_date", param_name = "param_name",
  analysis_result = "analysis_result", default_unit = "default_unit",
  magnesium = "Magnesium, dissolved", calcium = "Calcium, dissolved",
  sodium = "Sodium, dissolved", potassium = "Potassium, dissolved",
  chloride = "Chloride, total", sulfate = "Sulfate, total",
  alkalinity = "Alkalinity, total (lab)",
  total_dissolved_solids = NULL)
}
\description{
Function to transform data to Stiff diagram coordinates
}
