% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_plot.R
\name{.ts_plot}
\alias{.ts_plot}
\title{Helper function for plotting time series of groundwater data}
\usage{
.ts_plot(df, x = "sample_date", y = "analysis_result",
  lt_measure = "lt_measure", facet_var = NULL, group_var = NULL,
  scale_y_trans = "identity", trend = NULL, background = NULL,
  limit1 = NULL, limit2 = NULL, short_name = FALSE, pnt = 3,
  ncol = NULL)
}
\description{
Helper function for plotting time series of groundwater data
}
