% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_dist.R
\name{est_dist}
\alias{est_dist}
\title{Function to return either normal, lognormal, or non-parametric 
distribution.}
\usage{
est_dist(df, alpha = 0.05, method = "sw", keep_object = FALSE)
}
\arguments{
\item{df}{groundwater data frame in tidy format}

\item{alpha}{alpha}

\item{method}{default is "sw"}

\item{choices}{default is c("norm, "lnorm")}
}
\description{
Function to return either normal, lognormal, or non-parametric 
distribution.
}
