% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_dist_censored.R
\name{est_dist_censored}
\alias{est_dist_censored}
\title{Function to return either normal, lognormal, gamma, or non-parametric 
distribution estimate}
\usage{
est_dist_censored(df, left_censored, alpha = 0.05, method = "sf",
  keep_object = FALSE)
}
\arguments{
\item{df}{groundwater data frame in tidy format}

\item{left_censored}{logical vector.}

\item{alpha}{alpha}

\item{method}{default is "sf"}
}
\description{
Function to return either normal, lognormal, gamma, or non-parametric 
distribution estimate
}
