% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{event_summary}
\alias{event_summary}
\title{Function to export summary table for a sampling event}
\usage{
event_summary(df, wells, params, start, end)
}
\arguments{
\item{df}{groundwater data frame}

\item{start}{start date}

\item{end}{end date}
}
\description{
Function to export summary table for a sampling event
}
