% Generated by roxygen2 (4.0.1): do not edit by hand
\name{export_manages}
\alias{export_manages}
\title{Function to export data from manages to different formats in excel}
\usage{
export_manages(df, file)
}
\arguments{
\item{df}{data frame}

\item{file}{full file path name with extension for export}
}
\description{
Function to export data from manages to different formats in excel
}

