% Generated by roxygen2 (4.0.1): do not edit by hand
\name{export_OEPA}
\alias{export_OEPA}
\title{Function to export data from manages to excel format required by OEPA}
\usage{
export_OEPA(df, wells, constituents, file, plant, export_date = date())
}
\arguments{
\item{df}{data frame}

\item{wells}{list of wells to be exported}

\item{constituents}{list of constituents to be exported}

\item{file}{full file path name with extension for export}
}
\description{
Function to export data from manages to excel format required by OEPA
}

