% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_extract.R
\name{extract_function_params}
\alias{extract_function_params}
\title{Extract parameter information from a function}
\usage{
extract_function_params(fun_call)
}
\arguments{
\item{fun_call}{text string for function call for which parameter information should be returned.}
}
\value{
data.frame
}
\description{
Extract parameter information from a function
}
\examples{
extract_function_params("dplyr::count")
}
