% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_extract.R
\name{extract_package_info}
\alias{extract_package_info}
\title{Extract package information}
\usage{
extract_package_info(packages)
}
\arguments{
\item{packages}{a package name or vector of package names}
}
\value{
a tibble
}
\description{
Extract information about an R package, including version, source,
objects, and dependencies.
}
\examples{
extract_package_info(packages = "ggplot2")
extract_package_info(packages = c("ggplot2", "dplyr"))
}
