% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_extract.R
\name{extract_package_tree}
\alias{extract_package_tree}
\title{Extract entire dependency network for package or vector of packages}
\usage{
extract_package_tree(packages)
}
\arguments{
\item{packages}{package name or vector of package names}
}
\value{
a tibble
}
\description{
Extract entire dependency network for package or vector of packages
}
\examples{
extract_package_tree(packages = "ggplot2")
extract_package_tree(packages = c("ggplot2", "dplyr"))
}
