% Generated by roxygen2 (4.0.2): do not edit by hand
\name{from_csv}
\alias{from_csv}
\title{function to read data in csv format and convert date to POSIXct with lubridate}
\usage{
from_csv(path, date_format = "mdy")
}
\arguments{
\item{path}{path to the csv file of groundwater data in the format
with column names
location_id, param_name, default_unit, lt_measure, analysis_result}

\item{date_format}{date format as either mdy, or ymd passed to lubridate}
}
\description{
function to read data in csv format and convert date to POSIXct with lubridate
}

