% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_data.R
\name{from_excel}
\alias{from_excel}
\title{function to read data in excel format}
\usage{
from_excel(path, sheet = "Sheet1")
}
\arguments{
\item{path}{path to the excel  file of groundwater data in the format
 with column names
location_id, param_name, default_unit, lt_measure, analysis_result}

\item{sheet}{sheet name in spreadsheet}
}
\description{
function to read data in excel format
}
