\name{get_analytes}
\alias{get_analytes}
\title{Function to return a list of all the analytes}
\usage{
get_analytes(df)
}
\arguments{
  \item{df}{data frame of groundwater data in the format
  with column names location_id, param_name, default_unit,
  lt_measure, analysis_result}
}
\description{
Function to return a list of all the analytes
}

