% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mcl.R
\name{get_mcl_params}
\alias{get_mcl_params}
\title{Function to find EPA primary and seconday water constituents}
\usage{
get_mcl_params(df, type = "all")
}
\arguments{
\item{df}{dataframe of groundwater data}

\item{type}{can be all, primary, or secondary}
}
\description{
Function to find EPA primary and seconday water constituents
}
