\title{
Path to the current R script
  Chemins vers le script actif
}\name{get_path}
\alias{get_script_path}
alias{get_script_file_path}
\alias{get_file_path}
\description{
  \code{get_script_path()} get the full path to the current R scipt. \cr
  \code{get_script_file_path()} get the full path to folder where the current R scipt is saved. \cr
  \code{get_file_path()} get the path to folder where the current R scipt is saved.
}\usage{
  get_script_path()
  get_script_file_path()
  get_file_path()
}\value{
 A character string
}\author{
  Hugo Marthinet
}\examples{
## No arg
get_script_path()
get_script_file_path()
get_file_path()
}
