\name{get_well_names}
\alias{get_well_names}
\title{Function to return a list of all the location IDs}
\usage{
get_well_names(df)
}
\arguments{
  \item{df}{data frame of groundwater data in the format
  with column names location_id, param_name, default_unit,
  lt_measure, analysis_result}
}
\description{
Function to return a list of all the location IDs
}

