% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_wells}
\alias{get_wells}
\title{Function to return a list of all the location IDs}
\usage{
get_wells(df, location_id)
}
\arguments{
\item{df}{data frame of groundwater data in the format with column names}

\item{location_id}{column for the sample location id}
}
\description{
Function to return a list of all the location IDs
}
