% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem.R
\name{golem}
\alias{golem}
\title{golem}
\usage{
golem(ip)
}
\arguments{
\item{ip}{A vector one or more IP addresses}
}
\value{
A data frame with geolocation data.
}
\description{
Retrieves geolocation data for IP addresses from http://freegeoip.live.
}
\details{
\code{golem()} will retrieve the country, region, city, zip, time code, latitude, longitude, and metro code for IP addresses you give it.

Requires: \code{stringi} and \code{magrittr} (pipe)
}
\examples{
ips <- c('68.38.164.93', '192.198.85.59', '73.32.208.180', '2604:3d08:1580:2d0:1cd4:dafd:f848:4fb9')
ip_dat <- golem(ips)

}
