\name{groundwater_summary}
\alias{groundwater_summary}
\title{Function to summarize the number of samples, mean, sd, and percentage of
non-detects. This is useful for calculating the upper prediction limit.}
\usage{
groundwater_summary(df)
}
\arguments{
  \item{df}{data frame of groundwater monitoring network
  data}
}
\description{
Function to summarize the number of samples, mean, sd, and
percentage of non-detects. This is useful for calculating
the upper prediction limit.
}

