% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{grubbs_flag}
\alias{grubbs_flag}
\title{Function to flag outliers from Grubb's test for outliers}
\usage{
grubbs_flag(x, ...)
}
\arguments{
\item{x}{column of analysis results}

\item{...}{other arguments passed to grubbs test}
}
\description{
Function to flag outliers from Grubb's test for outliers
}

