% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{grubbs_test}
\alias{grubbs_test}
\title{Grubb's test for outliers}
\usage{
grubbs_test(df, x = "analysis_result", type = 10, opposite = FALSE,
  two_sided = FALSE, group_by_location = FALSE,
  keep_data_object = FALSE)
}
\arguments{
\item{df}{data frame}

\item{x}{column of analysis results}

\item{type}{integer value indicating test variant. 10 is a test for one
outlier (side is detected automatically and can be reversed by opposite
parameter). 11 is a test for two outliers on opposite tails, 20 is test
for two outliers in one tail. Default is 10}

\item{opposite}{a logical indicating whether you want to check not the
value with largest difference from the mean, but opposite (lowest, if
most suspicious is highest etc.). Default is FALSE}

\item{two_sided}{Logical value indicating if there is a need to treat
this test as two-sided. Default is FALSE}

\item{group_by_location}{TRUE/FALSE default is FALSE}

\item{keep_data_object}{TRUE/FALSE default is FALSE}
}
\description{
Grubb's test for outliers
}
