% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gw_boxplot}
\alias{gw_boxplot}
\title{Function to plot boxplots for groundwater data}
\usage{
gw_boxplot(df, facet_by = "location_id", name = NULL, coord_flip = FALSE)
}
\arguments{
\item{df}{groundwater data}

\item{facet_by}{parameter to group data by}

\item{name}{If name = "short" the analyte will be abbreviated}

\item{flip_coords}{If TRUE the axes are flipped}
}
\description{
Function to plot boxplots for groundwater data
}

