% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gw_data}
\alias{gw_data}
\title{Example groundwater quality data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5878 rows and 6 columns.}
\usage{
gw_data
}
\description{
Example groundwater quality data
}
\keyword{datasets}
