% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gw_gof}
\alias{gw_gof}
\title{Function to check for normal, lognormal, gamma, or non-parametric distribution}
\usage{
gw_gof(df, nd_percent = 10, non_par_perc = 50)
}
\arguments{
\item{df}{data frame of groundwater data}
}
\description{
Function to check for normal, lognormal, gamma, or non-parametric distribution
}

