% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gw_summary}
\alias{gw_summary}
\title{Function to summarize the number of samples, mean, sd, and percentage of
non-detects. This is useful for calculating the upper prediction limit.}
\usage{
gw_summary(df, bkgd_start, bkgd_end)
}
\arguments{
\item{df}{data frame of groundwater monitoring network data}
}
\description{
Function to summarize the number of samples, mean, sd, and percentage of
non-detects. This is useful for calculating the upper prediction limit.
}

