% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gw_ts_plot}
\alias{gw_ts_plot}
\title{Function to plot time series by location id}
\usage{
gw_ts_plot(df, facet_by = "location_id", back_date = NULL,
  comp_date = NULL, limit1 = NULL, limit2 = NULL, name = NULL,
  pnt = 3, ncol = NULL)
}
\arguments{
\item{df}{groundwater data}

\item{back_date}{dates for background date range}

\item{comp_date}{dates for compliance date range}

\item{limit1}{horizontal line 1}

\item{limit2}{horizontal line 2}

\item{name}{If name = "short" the analyte will be abbreviated}

\item{pnt}{size of points on time series plots}

\item{ncol}{number of columns}
}
\description{
Function to plot time series by location id
}

