% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{include_dup}
\alias{include_dup}
\title{function to include duplicate samples}
\usage{
include_dup(df, wells)
}
\arguments{
\item{df}{groundwater data frame}
}
\description{
function to include duplicate samples
}
