% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ind_by_loc}
\alias{ind_by_loc}
\title{This function plots groundwater data by location and faceted by parameters.
Other variables like shaded regions for background and compliance date ranges,
horizontal lines are for Groundwater Protection Standards can be passed to
the funcion.}
\usage{
ind_by_loc(df, ...)
}
\arguments{
\item{df}{data frame in long format containing groundwater monitoring data
assumes column names of location_id, param_name, analysis_result,
default_unit, lt_measure, sample_date. Dates for sample_date column
 must be in as.POSIXct format}

\item{back_date}{vector of start and end of background dates. Use as.POSIXct()}

\item{comp_date}{vector of start and end of compliance dates. Use as.POSIXct()}

\item{limit1}{Plots a vertical line for limit}

\item{limit2}{Plots a vertical line for a second limit}
}
\description{
This function plots groundwater data by location and faceted by parameters.
Other variables like shaded regions for background and compliance date ranges,
horizontal lines are for Groundwater Protection Standards can be passed to
the funcion.
}

