% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ind_by_param_grid}
\alias{ind_by_param_grid}
\title{Function to plot time series by parameter}
\usage{
ind_by_param_grid(df, back_date = NULL, comp_date = NULL, limit1 = NULL,
  limit2 = NULL, name = NULL, pnt = 3, ncol = 1)
}
\arguments{
\item{df}{groundwater data}

\item{back_date}{dates for background date range}

\item{comp_date}{dates for compliance date range}

\item{limit1}{horizontal line 1}

\item{limit2}{horizontal line 2}

\item{name}{If name = "short" the analyte will be abbreviated}

\item{pnt}{size of points on time series plots}

\item{ncol}{number of columns}
}
\description{
Function to plot time series by parameter
}

