% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{inline}
\alias{inline}
\title{A function ti display shiny elements inline}
\usage{
inline(x, m = 10, va = "top")
}
\arguments{
\item{x}{the elements to display}

\item{m}{margin size (in px)}

\item{va}{vertical alignement (top, middle, bottom)}
}
\value{
the element to display in a html div tag with the relevant css style
}
\description{
A function ti display shiny elements inline
}
\examples{
\dontrun{
inline(textInput("text", label = "Enter text", value = ""))
}
}
