% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{insert_fa}
\alias{insert_fa}
\title{Function to insert fontawesom icon in Rmd rendered with weasydoc}
\usage{
insert_fa(x)
}
\arguments{
\item{x}{string. Name of the fontawesome icon}
}
\value{
a html code as string
}
\description{
Function to insert fontawesom icon in Rmd rendered with weasydoc
}
\examples{
insert_fa("r-project")
}
