% Generated by roxygen2 (4.0.2): do not edit by hand
\name{intra_pred_int}
\alias{intra_pred_int}
\title{Function to calculate intrawell prediction interval}
\usage{
intra_pred_int(df, analysis_result, wells, params, bkgd_dates, comp_dates,
  SWFPR = 0.1, ...)
}
\arguments{
\item{df}{data frame of groundwater data}

\item{analysis_result}{name of column containing analysis results}

\item{wells}{vector of wells to be included}

\item{params}{vector of constituents to be included}

\item{bkgd_dates}{background data date range}

\item{SWFPR}{Site-Wide False Positive Rate}
}
\description{
Function to calculate intrawell prediction interval
}

