% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is_lognormal}
\alias{is_lognormal}
\title{Function to check if groundwater data has a Lognormal distribution}
\usage{
is_lognormal(x)
}
\arguments{
\item{x}{analysis_result of groundwater data}
}
\description{
Function to check if groundwater data has a Lognormal distribution
}

