% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality_censored.R
\name{is_lognormal_censored}
\alias{is_lognormal_censored}
\title{Helper function to check if censored groundwater data has a
Lognormal distribution}
\usage{
is_lognormal_censored(x, censored, p = 0.01, ...)
}
\description{
Helper function to check if censored groundwater data has a
Lognormal distribution
}
