% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{is_normal}
\alias{is_normal}
\title{Helper function to check if data has a Normal distribution}
\usage{
is_normal(x, p = 0.01)
}
\description{
Helper function to check if data has a Normal distribution
}
