% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is_normal_censored}
\alias{is_normal_censored}
\title{Function to check if censored groundwater data has a Normal distribution}
\usage{
is_normal_censored(x, censored, ...)
}
\arguments{
\item{x}{analysis_result of groundwater data}

\item{censored}{logical vector of censored values}
}
\description{
Function to check if censored groundwater data has a Normal distribution
}

