% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{join_lt}
\alias{join_lt}
\title{Function to join columns of lt_measure and sample results}
\usage{
join_lt(df, lt_measure, analysis_result)
}
\arguments{
\item{df}{groundwater data frame}

\item{lt_measure}{column name for non-detect symbol}

\item{analysis_result}{column name for analysis results}
}
\description{
Function to join columns of lt_measure and sample results
}
