% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{kendall_trend}
\alias{kendall_trend}
\title{Kendall Trend}
\usage{
kendall_trend(df, conf.level = 0.99)
}
\arguments{
\item{df}{dataframe of groundwater data}

\item{conf.level}{confidence level between 0 and 1}
}
\description{
Kendall Trend
}
