\title{
  Chargement de packages
}\name{libraryCESPA}\alias{libraryCESPA}\description{
  Charge les packages developpes au CESPA : \code{\link{manageR}}, \code{\link{epiplus}}, \code{\link{MDBprod}} et \code{\link{lazyMe}}
}\usage{
  libraryCESPA()
}\author{
  Hugo Marthinet
}\examples{
 libraryCESPA()
}
