\title{
  Marthinet's packages loading adn installing
}\name{libraryMarthinet}
\alias{libraryMarthinet}
\alias{installMarthinet}
\description{
  Load and/or install all availaible package on Hugo Marthinet's github. That would be \code{dataDesc}, \code{epiplus}, \code{surveyscores}, \code{sampleSize}, \code{lazyMe} and \code{manageR}.
}\usage{
installMarthinet()
libraryMarthinet()
}\author{
  Hugo Marthinet
}\examples{

}
