% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lim_sim}
\alias{lim_sim}
\title{Function to calculate intrawell prediction interval}
\usage{
lim_sim(x, nd_percent = 10, non_par_perc = 50, nd_method = "ROS")
}
\arguments{
\item{df}{data frame of groundwater data}

\item{wells}{vector of wells to be included}

\item{params}{vector of constituents to be included}

\item{bkgd_dates}{background data date range}

\item{SWFPR}{Site-Wide False Positive Rate}
}
\description{
Function to calculate intrawell prediction interval
}

