% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_edit_settingsInput.R
\name{mod_edit_settingsInput}
\alias{mod_edit_settingsInput}
\alias{mod_edit_settings}
\title{mod_edit_settingsInput and mod_edit_settings}
\usage{
mod_edit_settingsInput(id)

mod_edit_settings(input, output, session, settingsdata, package)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{settingsdata}{the reactive xml settings file}

\item{package}{package where the xml config file is}
}
\value{
the modified reactive xml settings file
}
\description{
A shiny module to manage a settings file stored as xml in the www
folder of the package

mod_edit_settings server function
}
