% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loading_optionsInput.R
\name{mod_loading_optionsInput}
\alias{mod_loading_optionsInput}
\alias{mod_loading_options}
\title{mod_loading_optionsInput and mod_loading_options}
\usage{
mod_loading_optionsInput(id)

mod_loading_options(input, output, session, path, filename,
  coltypes = cols(.default = col_character()))
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{inernal}

\item{path}{reactive string value. The path to the directory where the file is}

\item{filename}{reactive string value. Name and extension of the file to read}

\item{coltypes}{see \code{col_types} argument in \code{read_delim}}
}
\description{
A shiny module that read a delimited file (with \code{read_delim}) from a path and a filename. It is more a internal module than a module that works on itself.

mod_loading_options server function
}
\examples{
library(shiny)
if (interactive()) {
ui <- fluidPage(
  h4("Choose a path"),
  textInput("path", value = getwd(), label = NULL),
  h4("Choose a filename (do not forget the extension)"),
  textInput("filename", value = "", label = NULL),
  h4("The module goes from here............................"),
  mod_loading_optionsInput("loadfile"),
  h4("..........................................to here"),
  h4("You can display the result"),
  tableOutput("data")
)

server <- function(input, output, session) {
  data <- callModule(mod_loading_options,"loadfile", path = reactive(input$path), 
  filename = reactive(input$filename))
  output$data <- renderTable(data())
}

shinyApp(ui, server)
}

}
