% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wd_loadUI.R
\name{mod_wd_loadUI}
\alias{mod_wd_loadUI}
\alias{mod_wd_load}
\title{mod_wd_loadUI and mod_wd_load}
\usage{
mod_wd_loadUI(id)

mod_wd_load(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny module that load and display a path to a working directory
}
\examples{
library(shiny)
library(shinyFiles)
if (interactive()){
ui <- fluidPage(
  mod_wd_loadUI("wd")
)

server <- function(input, output, session) {
  path <- callModule(mod_wd_load,"wd")
}

shinyApp(ui, server)
}

}
