% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multi_by_loc}
\alias{multi_by_loc}
\title{This function plots groundwater data by location with all parameters
on one plot. Shaded regions for background and compliance date ranges,
horizontal lines for Groundwater Protection Standards can be passed to the
function as well}
\usage{
multi_by_loc(df, ...)
}
\arguments{
\item{df}{data frame in long format containing groundwater monitoring data
assumes column names of location_id, param_name, analysis_result,
default_unit, lt_measure, sample_date. Dates for sample_date column
 must be in as.POSIXct format}

\item{back_date}{vector of start and end of background dates. Use as.POSIXct()}

\item{comp_date}{vector of start and end of compliance dates. Use as.POSIXct()}

\item{limit1}{plots a horizontal line for a limit e.g. EPA MCL}

\item{limit2}{plots a second limit}
}
\description{
This function plots groundwater data by location with all parameters
on one plot. Shaded regions for background and compliance date ranges,
horizontal lines for Groundwater Protection Standards can be passed to the
function as well
}

