% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multi_by_param}
\alias{multi_by_param}
\title{This function plots groundwater data faceted by parameters with shaded
regions for background and compliance date ranges.
Horizontal lines are plotted for Groundwater Protection Standards.}
\usage{
multi_by_param(df, ...)
}
\arguments{
\item{df}{data frame in long format containing groundwater monitoring data
assumes column names of location_id, param_name, analysis_result,
default_unit, lt_measure, sample_date. Dates for sample_date column
 must be in as.POSIXct format}

\item{facet_col}{the column to facet the plot by}

\item{back_date}{vector of start and end of background dates.}

\item{comp_date}{vector of start and end of compliance dates.}

\item{limits}{column vector of limits e.g. c("EPA_Limits", "DMR_limits")}
}
\description{
This function plots groundwater data faceted by parameters with shaded
regions for background and compliance date ranges.
Horizontal lines are plotted for Groundwater Protection Standards.
}

