% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multi_gw_boxplot}
\alias{multi_gw_boxplot}
\title{Function to plot multiple boxplots for groundwater data}
\usage{
multi_gw_boxplot(df, facet_by = "location_id", ...)
}
\arguments{
\item{df}{groundwater data}

\item{facet_by}{parameter to group data by}

\item{name}{If name = "short" the analyte will be abbreviated}

\item{flip_coords}{If TRUE the axes are flipped}
}
\description{
Function to plot multiple boxplots for groundwater data
}

