% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multi_gw_ts_plot}
\alias{multi_gw_ts_plot}
\title{This function plots multiple groundwater data time series by location,
or constituent. Other variables from \code{\link{gw_ts_plot}} can be passed
to the funcion.}
\usage{
multi_gw_ts_plot(df, facet_by = "location_id", ...)
}
\arguments{
\item{df}{data frame in long format containing groundwater monitoring data
assumes column names of location_id, param_name, analysis_result,
default_unit, lt_measure, sample_date. Dates for sample_date column
 must be in as.POSIXct format}

\item{facet_by}{parameter to group by. Can be either location_id,
or param_name. Defaults to location_id.}
}
\description{
This function plots multiple groundwater data time series by location,
or constituent. Other variables from \code{\link{gw_ts_plot}} can be passed
to the funcion.
}

