% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_helper.R
\name{nth_date}
\alias{nth_date}
\title{Function to find the date for next n samples}
\usage{
nth_date(df, start, n = 4, sample_date)
}
\arguments{
\item{df}{df dataframe of groundwater data in tidy format}

\item{start}{start the date to start from}

\item{n}{n the next n samples to get the date for}

\item{sample_date}{sample_date the column for the sampling dates}
}
\description{
Function to find the date for next n samples
}
