% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ohio_data}
\alias{ohio_data}
\title{Groundwater quality data for the state of Ohio}
\format{A data frame with 213845 rows and 47 variables:
\describe{
  \item{well_id}{well_id, Site Identifier}
  \item{result_value}{analytical result}
  \item{analyte_count}{}
  \item{analyte_group}{}
  \item{analyte_id}{}
  \item{casing_length}{}
  \item{charge_balance_error}{}
  \item{chemistry_sheet_id}{}
  \item{collection_timestamp}{}
  \item{collector_gwqcp_person_id}{}
  \item{comb_result}{}
  \item{county}{}
  \item{detect_limit_unit}{}
  \item{detect_limit_value}{}
  \item{detected_count}{}
  \item{district}{}
  \item{field_comments}{}
  \item{field_qc_type}{}
  \item{geo_age}{}
  \item{geo_setting}{}
  \item{is_filtered}{}
  \item{lab_comments}{}
  \item{lab_comments.1}{}
  \item{latitude}{}
  \item{lith_open_section}{}
  \item{local_name}{}
  \item{longitude}{}
  \item{major_lithology}{}
  \item{matrix_type}{}
  \item{method}{}
  \item{name}{}
  \item{name.1}{}
  \item{original_sheet_id}{}
  \item{parameter}{}
  \item{pws_id}{}
  \item{remark_code}{}
  \item{result_unit}{}
  \item{sample_id}{}
  \item{sample_type}{}
  \item{sampling_status}{}
  \item{site_id}{}
  \item{status}{}
  \item{user_comments}{}
  \item{well_depth}{}
  \item{well_id_code}{}
  \item{well_log_num}{}
  \item{well_number}{}
}}
\source{
\url{http://epa.ohio.gov/ddagw/gwqcp.aspx}
}
\usage{
ohio_data
}
\description{
A dataset containing groundwater quality data for the state of Ohio.
}
\keyword{datasets}
