% Generated by roxygen2 (4.0.2): do not edit by hand
\name{percent_lt}
\alias{percent_lt}
\title{calculate the percentage of left censored data}
\usage{
percent_lt(lt)
}
\arguments{
\item{lt}{column of data less than detection limit.}

\item{df}{data frame of groundwater monitoring data in long format}
}
\description{
calculate the percentage of left censored data
}

