% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper_plots.R
\name{piper_plot}
\alias{piper_plot}
\title{Create a piper plot}
\usage{
piper_plot(df, location_id = "LOCATION_ID", sample_date = "SAMPLE_DATE",
  x_cation = "Calcium, dissolved", y_cation = "Magnesium, dissolved",
  z_cation = c("Sodium, dissolved", "Potassium, dissolved"),
  x_anion = c("Chloride, total", "Fluoride, total"),
  y_anion = "Alkalinity, total (lab)", z_anion = "Sulfate, total",
  pnt = 3, transparency = 0.2, TDS = FALSE, title = NULL)
}
\arguments{
\item{df}{data frame of water quality data in tidy format}

\item{location_id}{column for sample location}

\item{sample_date}{column for sample date}

\item{x_cation}{default is Calcium, dissolved}

\item{y_cation}{default is Magnesium, dissolved}

\item{z_cation}{default is Sodium, dissolved + Potassium, dissolved}

\item{x_anion}{default is Chloride, total + Fluoride, total}

\item{y_anion}{default is Alkalinity, total (lab)}

\item{z_anion}{default is Sulfate, total}

\item{TDS}{Scale plot by Total Dissolved Solids, default = FALSE}

\item{title}{Title for plot, default = NULL}
}
\description{
Create a piper plot
}
