% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochemistry_plots.R
\name{piper_time_html}
\alias{piper_time_html}
\title{Function to create an aminated Piper plot and save to html}
\usage{
piper_time_html(df, TDS = FALSE)
}
\arguments{
\item{df}{data frame of groundwater data transformed using 
\code{\link{transform_piper_data}}}

\item{TDS}{Scale by Total Dissolved Solids}
}
\description{
Function to create an aminated Piper plot and save to html
}
