% Generated by roxygen2 (4.0.2): do not edit by hand
\name{piper_time_plot}
\alias{piper_time_plot}
\title{Function to created an animated Piper plot through time
using the animation package}
\usage{
piper_time_plot(df, TDS = FALSE, title)
}
\arguments{
\item{df}{data frame of groundwater data transformed using
\code{\link{transform_piper_data}}}

\item{TDS}{Scale by Total Dissolved Solids}
}
\description{
Function to created an animated Piper plot through time
using the animation package
}

