% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_extract.R
\name{plot_inventory}
\alias{plot_inventory}
\title{Plot dependency tree}
\usage{
plot_inventory(inventory_object)
}
\arguments{
\item{inventory_object}{a tibble exported from the `take_inventory` function}
}
\value{
a ggplot visualization of the given dependency tree
}
\description{
Plot dependency tree
}
\examples{
ggplot2_inventory <- take_inventory(packages = "ggplot2")
plot_inventory(inventory_object = ggplot2_inventory)
}
