\name{plot_piper}
\alias{plot_piper}
\title{Function to plot points from transform_piper_data() onto base ggplot_piper
background. Points are sized by Total Dissolved Solids}
\usage{
plot_piper(df)
}
\arguments{
  \item{df}{data frame of groundwater data transformed into
  piper coordinates using
  \code{\link{transform_piper_data}}}
}
\description{
Function to plot points from transform_piper_data() onto
base ggplot_piper background. Points are sized by Total
Dissolved Solids
}

