% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_extract.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot dependency tree}
\usage{
plot_tree(tree_object)
}
\arguments{
\item{tree_object}{a tibble exported from the `extract_package_tree` function}
}
\value{
a ggplot visualization of the given dependency tree
}
\description{
Plot dependency tree
}
\examples{
ggplot2_tree <- extract_package_tree(packages = "ggplot2")
plot_tree(tree_object = ggplot2_tree)
}
