% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_int.R
\name{pred_int_sim}
\alias{pred_int_sim}
\title{Function to calculate simultaneous prediction interval}
\usage{
pred_int_sim(df, analysis_result = "analysis_result",
  distribution = "distribution", n_mean = 1, k = 1, m = 1, r = 1,
  rule = "k.of.m", pi_type = "upper", conf_level = 0.95)
}
\arguments{
\item{df}{df data frame of groundwater data in tidy format}

\item{analysis_result}{the analysis result column}

\item{distribution}{the distribution column}

\item{k}{k positive integer specifying the number of future observations or
averages the prediction interval should contain with confidence level
conf.level. The default value is k=1.}

\item{m}{m if m = 1, then same results as predInt}

\item{r}{r}

\item{rule}{rule "k.of.m"}

\item{pi_type}{character string indicating what kind of prediction interval
to compute. The possible values are pi_type="upper" (the default),
and pi_type="lower".}

\item{n.mean}{n.mean positive integer specifying the sample size associated
with the future averages.
The default value is n.mean=1 (i.e., individual observations).
Note that all future averages must be based on the same sample size.}

\item{conf.level}{a scalar between 0 and 1 indicating the confidence level
of the prediction interval. The default value is conf.level=0.95}
}
\description{
Function to calculate simultaneous prediction interval
}
