% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_functions.R
\name{process_annexes}
\alias{process_annexes}
\title{Internal function to process annexes. Do not use outside}
\usage{
process_annexes(file)
}
\arguments{
\item{file}{character. Path of the annexes file}
}
\description{
Internal function to process annexes. Do not use outside
}
