% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_functions.R
\name{process_bank}
\alias{process_bank}
\title{Internal function to process the bank part in Rmd template. Do not use outside}
\usage{
process_bank(info, bankinfo, config)
}
\arguments{
\item{info}{info list from YAML params}

\item{bankinfo}{bank list from YAML params}

\item{config}{config list from YAML params}
}
\description{
Internal function to process the bank part in Rmd template. Do not use outside
}
