% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_functions.R
\name{process_boxheader}
\alias{process_boxheader}
\title{Internal function to process the boxheader part in Rmd template. Do not use outside}
\usage{
process_boxheader(info, which)
}
\arguments{
\item{info}{list of info from the YAML params}

\item{which}{either 1, 2 or 3 depending of the boxheader targeted}
}
\description{
Internal function to process the boxheader part in Rmd template. Do not use outside
}
