% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_functions.R
\name{process_client}
\alias{process_client}
\title{Internal function to process the client and billing parts in Rmd template. Do not use outside}
\usage{
process_client(client, billing, info)
}
\arguments{
\item{client}{client list from YAML params}

\item{billing}{billing list from YAML params}

\item{info}{info list from YAML params}
}
\description{
Internal function to process the client and billing parts in Rmd template. Do not use outside
}
