% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_functions.R
\name{process_comment}
\alias{process_comment}
\title{Internal function to process the comment and billing parts in Rmd template. Do not use outside}
\usage{
process_comment(comment, info)
}
\arguments{
\item{comment}{comment from YAML params}

\item{info}{info list from YAML params}
}
\description{
Internal function to process the comment and billing parts in Rmd template. Do not use outside
}
