% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_functions.R
\name{process_config}
\alias{process_config}
\title{Internal function to process the config part in Rmd template. Do not use outside}
\usage{
process_config(config)
}
\arguments{
\item{config}{list of config from the YAML params}
}
\description{
Internal function to process the config part in Rmd template. Do not use outside
}
