% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_functions.R
\name{process_logo}
\alias{process_logo}
\title{Internal function to process the logo part in Rmd template. Do not use outside}
\usage{
process_logo(logo)
}
\arguments{
\item{logo}{list of logo from the YAML params}
}
\description{
Internal function to process the logo part in Rmd template. Do not use outside
}
