% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_functions.R
\name{process_name}
\alias{process_name}
\title{Internal function to process the company name part in Rmd template. Do not use outside}
\usage{
process_name(name)
}
\arguments{
\item{name}{name of company from the YAML params}
}
\description{
Internal function to process the company name part in Rmd template. Do not use outside
}
