% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_functions.R
\name{process_services}
\alias{process_services}
\title{Internal function to process the services part in Rmd template. Do not use outside}
\usage{
process_services(services)
}
\arguments{
\item{services}{services list from YAML params}
}
\description{
Internal function to process the services part in Rmd template. Do not use outside
}
