% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_functions.R
\name{process_total}
\alias{process_total}
\title{Internal function to process the total part in Rmd template. Do not use outside}
\usage{
process_total(services)
}
\arguments{
\item{services}{services list from YAML params}
}
\description{
Internal function to process the total part in Rmd template. Do not use outside
}
