% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_functions.R
\name{process_tva}
\alias{process_tva}
\title{Internal function to process the tva part in Rmd template. Do not use outside}
\usage{
process_tva(services)
}
\arguments{
\item{services}{services list from YAML params}
}
\description{
Internal function to process the tva part in Rmd template. Do not use outside
}
