% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_data.R
\name{read_manages3_spatial}
\alias{read_manages3_spatial}
\title{function to connect to MANAGES database and read spatial data}
\usage{
read_manages3_spatial(manages_path)
}
\arguments{
\item{manages_path}{path to MANAGES database}
}
\description{
function to connect to MANAGES database and read spatial data
}
