% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_data.R
\name{read_manages4}
\alias{read_manages4}
\title{Function to read MANAGES 4.0 database}
\usage{
read_manages4(driver = "SQL Server", server, database, site)
}
\arguments{
\item{driver}{default is "SQL Server"}

\item{server}{server name}

\item{database}{database name}

\item{site}{list of sites}
}
\description{
Function to read MANAGES 4.0 database
}
