% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{remove_dup}
\alias{remove_dup}
\title{Function to remove duplicate samples
Example: If you have wells named MW-1 and another named MW-1 Duplicate
this function will remove the MW-1 Duplicate sample}
\usage{
remove_dup(df)
}
\arguments{
\item{df}{groundwater data frame}
}
\description{
Function to remove duplicate samples
Example: If you have wells named MW-1 and another named MW-1 Duplicate
this function will remove the MW-1 Duplicate sample
}
