% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{replace_missing}
\alias{replace_missing}
\title{Function to replace missing values}
\usage{
replace_missing(df, value)
}
\arguments{
\item{df}{groundwater data frame}

\item{value}{value to replace missing value with}
}
\description{
Function to replace missing values
}
