% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{rosner_flag}
\alias{rosner_flag}
\title{Function to flag outliers from Rosner's test for outliers}
\usage{
rosner_flag(df, x = "analysis_result", replace = NULL, ...)
}
\arguments{
\item{df}{data.frame of groundwater data}

\item{x}{column of analysis results}

\item{replace}{value to replace outliers with}

\item{...}{other arguments passed to rosner test}
}
\description{
Function to flag outliers from Rosner's test for outliers
}

