% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{rosner_test}
\alias{rosner_test}
\title{Rosner's test for outliers}
\usage{
rosner_test(df, x = "analysis_result", k = 3, alpha = 0.05, warn = TRUE,
  group_by_location = FALSE, keep_data_object = FALSE)
}
\arguments{
\item{df}{data.frame of groundwater data}

\item{x}{column of analysis results}

\item{k}{number of suspected outliers. Default is 3.}

\item{alpha}{numerical scalar bewteen 0 and 1 indicating the Type I error}

\item{warn}{logical scalar indicating whether to issue a warning
(warn=TRUE; the default) when the number of non-missing, finite values
in x and the value of k are such that the assumed Type I error level might
not be maintained.}

\item{group_by_location}{TRUE/FALSE}

\item{keep_data_object}{TRUE/FALSE to return entire gofOutlier object}
}
\description{
Rosner's test for outliers
}
