% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{sample_locations}
\alias{sample_locations}
\title{Function to return a list of all the location IDs}
\usage{
sample_locations(df, location_id)
}
\arguments{
\item{df}{data frame of groundwater data in the format with column names}

\item{location_id}{column for the sample location id}
}
\description{
Function to return a list of all the location IDs
}
