% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochemistry_plots.R
\name{schoeller_plot}
\alias{schoeller_plot}
\title{Function to plot data transformed to schoeller coordinates using
\code{\link{transform_schoeller}}}
\usage{
schoeller_plot(df, facet_by = NULL, title = NULL, lwt = 1)
}
\arguments{
\item{df}{dataframe of transformed groundwater data}

\item{facet_by}{parameter to facet plots by}

\item{title}{title of plot}

\item{lwt}{lineweight}
}
\description{
Function to plot data transformed to schoeller coordinates using
\code{\link{transform_schoeller}}
}
