% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_extract.R
\name{script_check}
\alias{script_check}
\title{Check an R script for presence of affected object changes}
\usage{
script_check(
  compare_object,
  script_filepath,
  is_R_script = T,
  include_specials = F,
  summary_file = ""
)
}
\arguments{
\item{compare_object}{Comparison object of two inventories as returned by `compare_inventory`.}

\item{script_filepath}{Filepath to a script of interest.}

\item{is_R_script}{Logical. Default is TRUE, which assumes the script is a .R file. If FALSE, applies `knitr::purl` to convert to a .R script prior to parsing.}

\item{include_specials}{Logical. Whether to include special characters as objects when searching script.}

\item{summary_file}{Character. Default is `""`, which prints resulting summary to the console. If a filepath to a .txt file is provided, summary will save to the file instead.}
}
\value{
Concatenated summary of whether script involves any affected objects from the inventory comparison object.
}
\description{
Check an R script for presence of affected object changes
}
\examples{
\dontrun{script_check(compare_object, script_filepath, summary_file = "script-check-results.txt")}
}
