% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_helper.R
\name{set_background}
\alias{set_background}
\title{Function to return a column of background data}
\usage{
set_background(df, sample_date, start, end)
}
\arguments{
\item{df}{dataframe of groundwater data in tidy format}

\item{sample_date}{the column for the sampling dates}

\item{start}{the start date for background}

\item{end}{the end date for background}
}
\description{
Function to return a column of background data
}
