% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{simplify_param_name}
\alias{simplify_param_name}
\title{Helper function to split parameter names}
\usage{
simplify_param_name(df, param_name)
}
\arguments{
\item{df}{data frame in long format}

\item{param_name}{parameter name column}
}
\description{
Helper function to split parameter names
}
