% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmax.R
\name{softmax}
\alias{softmax}
\title{softmax}
\usage{
softmax(x, digits = NULL)
}
\arguments{
\item{x}{A double or integer vector.}

\item{digits}{The number of digits to round the output probabilities to.}
}
\value{
A length-one vector.
}
\description{
Calculates the softmax function for mapping a set of real values to a probability distribution.
The maxima/um receives quite a bit of weight via softmax: Given the values \code{[4, 4, 4, 5]}, softmax
returns a probability of .475 for the final element and 0.175 for each of the first three.
}
\examples{
x <- c(3.5, 4.0, 6.5)
softmax(x)
softmax(x, digits = 4)

x <- c(NA, 3.5, 4.0, 6.5)
softmax(x, 4)

}
