% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat4DS_data.R
\name{stat4DS_data}
\alias{stat4DS_data}
\title{stat4DS_data}
\usage{
stat4DS_data(dat, preview = FALSE)
}
\arguments{
\item{dat}{A length-one character vector; the name of a data set or \code{'all'} to return names of all possible data sets.}

\item{preview}{If \code{TRUE}, only return the first few rows of the data set; default is \code{FALSE}.}
}
\value{
A data set if \code{dat} is a data set name; console output (and \code{NULL} invisibly) if \code{dat} is \code{'all'}.
}
\description{
Retrieve data sets used in \emph{Foundations of Statistics for Data Scientists With R and Python}
(Agresti & Kateri, 2021) for use as demo/test data.
}
\examples{
stat4DS_data('Florida', preview = TRUE)
stat4DS_data('all')

}
