% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochemistry_plots.R
\name{stiff_by_loc}
\alias{stiff_by_loc}
\title{Function to plot multiple stiff plots by location}
\usage{
stiff_by_loc(df, ...)
}
\arguments{
\item{df}{groundwater data frame}

\item{...}{arguments passed to \code{\link{stiff_plot}}}
}
\description{
Function to plot multiple stiff plots by location
}
