\name{stiff_time_html}
\alias{stiff_time_html}
\title{function to create an animated Stiff Diagram and save to html}
\usage{
stiff_time_html(df)
}
\arguments{
  \item{df}{data frame of groundwater data transformed
  using \code{\link{transform_stiff_data}}}
}
\description{
function to create an animated Stiff Diagram and save to
html
}

