% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiff_plot.R
\name{stiff_time_plot}
\alias{stiff_time_plot}
\title{Function to create an animated Stiff Diagram}
\usage{
stiff_time_plot(df, ...)
}
\arguments{
\item{df}{data frame of water quality data}

\item{...}{
  Arguments passed on to \code{\link[=stiff_plot]{stiff_plot}}
  \describe{
    \item{\code{location_id}}{column for sample locations}
    \item{\code{sample_date}}{column for sample date}
    \item{\code{param_name}}{column for constituent names}
    \item{\code{analysis_result}}{column for analysis result}
    \item{\code{default_unit}}{column for units. Assumes all units are in mg/L and
converts usign \code{\link{conc_to_meq}}}
    \item{\code{magnesium}}{name for magnesium}
    \item{\code{calcium}}{name for calcium}
    \item{\code{sodium}}{name for sodium}
    \item{\code{potassium}}{name for potassium}
    \item{\code{chloride}}{name for chloride}
    \item{\code{sulfate}}{name for sulfate}
    \item{\code{alkalinity}}{name for alkalinity}
    \item{\code{total_dissolved_solids}}{name for TDS to scale color of plot.
Default is NULL}
    \item{\code{group_var}}{parameter to separate plots by. Default is sample location}
    \item{\code{facet_var}}{parameter to facet the plot by. Default is sample date}
    \item{\code{lines}}{TRUE/FALSE plots lines}
    \item{\code{cex}}{multiplier to scale plot height}
  }}
}
\description{
Function to create an animated Stiff Diagram
}
