% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_extract.R
\name{take_inventory}
\alias{take_inventory}
\title{Take inventory of entire dependency network for package or vector of packages}
\usage{
take_inventory(packages)
}
\arguments{
\item{packages}{package name or vector of package names}
}
\value{
a tibble
}
\description{
Take inventory of entire dependency network for package or vector of packages
}
\examples{
take_inventory(packages = "ggplot2")
take_inventory(packages = c("ggplot2", "dplyr"))
}
