% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_person_regex.R
\docType{data}
\name{test_names}
\alias{test_names}
\title{Set of names for testing the efficacy of \code{non_person_regex}}
\format{
A data frame with 232 rows and 2 variables:
\describe{
  \item{\code{name}}{The name}
  \item{\code{type}}{The type of name, non-person or person}
}
}
\source{
Created by github.com/jacob-gg
}
\usage{
test_names
}
\description{
A data frame containing a mix of non-person names (e.g., "Downtown Books," "Supreme Vacuums, LLC") and individuals' names that can be used to test the efficacy of the regex pattern \code{non_person_regex}.
}
\keyword{datasets}
