% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{theil_sen_band}
\alias{theil_sen_band}
\title{Function to calculate the Theil-Sen confidence band}
\usage{
theil_sen_band(ANALYSIS_RESULT, SAMPLE_DATE, conf.level)
}
\description{
Function to calculate the Theil-Sen confidence band
}
