% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camel_snake.R
\name{to_camel}
\alias{to_camel}
\title{to_camel}
\usage{
to_camel(x, nums_to_camel = T)
}
\arguments{
\item{x}{A vector (character or coercible to character).}

\item{nums_to_camel}{Should numbers be cameled as well? ('z_1_a' --> 'z1A')}
}
\value{
A vector.
}
\description{
Converts a vector of strings (usually column names) from snake_case to camelCase. (Specifically, lower camel case.)
}
\details{
\code{to_camel()} will return the input with a warning unless there is at least one underscore not in the first position somewhere in the set of input strings.

E.g., \code{to_camel(c('aaa', 'bbb', '_ccc'))} will return \code{x} and a warning.

No dependencies.
}
\examples{
snakes <- c('rainbow_boa', 'blue_racer', 'dragon_snake',
'a_set_of_5_garter_snakes', 'a_larger_set_of_6_kingsnakes')
to_camel(snakes)

}
