% Generated by roxygen2 (4.0.2): do not edit by hand
\name{to_censored}
\alias{to_censored}
\title{Function to convert gwdata frame to censored data frame}
\usage{
to_censored(df)
}
\arguments{
\item{df}{data frame of groundwater data}
}
\description{
Function to convert gwdata frame to censored data frame
}

