% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{to_spatial}
\alias{to_spatial}
\title{function to convert data to sf object}
\usage{
to_spatial(df, crs)
}
\arguments{
\item{df}{data frame}

\item{crs}{coordinate reference system to convert the sf object to}
}
\description{
function to convert data to sf object
}
