% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{to_wide}
\alias{to_wide}
\title{Helper function to cast data frame to wide format}
\usage{
to_wide(df, lab_id = FALSE, join_lt = TRUE)
}
\arguments{
\item{df}{data frame in long format}

\item{lab_id}{logical to include lab_id. Default is FALSE.}

\item{join_lt}{logical to join the non-detect colum with analysis result.}
}
\description{
Helper function to cast data frame to wide format
}
\examples{
data(gw_data)

gw_data \%>\%
  fitler(param_name \%in\% c("Arsenic, dissolved", "Boron, dissolved")) \%>\%
  to_wide(., lab_id = TRUE)
}
