% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tol_int.R
\name{tol_int}
\alias{tol_int}
\title{Function to calculate tolernace intervals}
\usage{
tol_int(df, analysis_result = "analysis_result",
  distribution = "distribution", coverage = 0.95,
  cov_type = "content", ti_type = "two-sided", method = "exact")
}
\arguments{
\item{df}{df data frame of groundwater data in tidy format}

\item{analysis_result}{the analysis result column}

\item{distribution}{the distribution column}

\item{coverage}{a scalar between 0 and 1 indicating the desired coverage of
the tolerance interval. The default value is coverage=0.95.
If cov.type="expectation", this argument is ignored}

\item{cov_type}{character string specifying the coverage type for the
tolerance interval. The possible values are "content"
(β-content; the default), and "expectation" (β-expectation).}

\item{ti_type}{character string indicating what kind of tolerance interval
to compute. The possible values are ti_type="two-sided",
ti_type="upper", and ti_type="lower" (the default).}

\item{method}{for the case of a two-sided tolerance interval, a character
string specifying the method for constructing the tolerance interval.
This argument is ignored if ti.type="lower" or ti.type="upper".
The possible values are "exact" (the default) and
"wald.wolfowitz" (the Wald-Wolfowitz approximation).}

\item{conf_level}{a scalar between 0 and 1 indicating the confidence level
of the tolerance interval. The default value is conf.level=0.95}

\item{est_method}{for tolIntLnormAlt, a character string specifying the
method of estimating the mean and coefficient of variation.
This argument has no effect on the method of constructing the
tolerance interval. Possible values are "mvue" (minimum variance unbiased;
the default), "qmle" (quasi maximum likelihood), "mle" (maximum likelihood),
 "mme" (method of moments), and "mmue" (method of moments based on the
 unbiased estimate of variance).}
}
\description{
Function to calculate tolernace intervals
}
