% Generated by roxygen2 (4.0.1): do not edit by hand
\name{transform_piper_data}
\alias{transform_piper_data}
\title{function to transform data from get_plot_data() into x, y coordinates for
cation, anion and diamond of a Piper plot}
\usage{
transform_piper_data(df, Mg = "Magnesium, dissolved",
  Ca = "Calcium, dissolved", Na = "Sodium, dissolved",
  K = "Potassium, dissolved", Cl = "Chloride, total",
  SO4 = "Sulfate, total", CO3 = NULL, Alk = "Alkalinity, total (lab)",
  TDS = "Total Dissolved Solids", name = "location_id",
  date = "sample_date", units = NULL)
}
\arguments{
\item{df}{data frame of groundwater data}

\item{Mg}{Magnesium}

\item{Ca}{Calcium}

\item{Na}{Sodium}

\item{K}{Potassum}

\item{Cl}{Chloride}

\item{SO4}{Sulfate}

\item{HCO3}{Bicarbonate, or Alkalinity}

\item{TDS}{Total Dissolved Solids}

\item{name}{column of well names}

\item{date}{column of dates}

\item{units}{units, right now can handle mg/L and meq/L}
}
\description{
function to transform data from get_plot_data() into x, y coordinates for
cation, anion and diamond of a Piper plot
}
\keyword{diagram}
\keyword{piper}

