% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochemistry_plots.R
\name{transform_schoeller}
\alias{transform_schoeller}
\title{Function to transoform data from \code{\link{get_major_ions}} to schoeller 
coordinates}
\usage{
transform_schoeller(df, Mg = "Magnesium, dissolved",
  Ca = "Calcium, dissolved", Na = "Sodium, dissolved",
  K = "Potassium, dissolved", Cl = "Chloride, total",
  SO4 = "Sulfate, total", HCO3 = "Alkalinity, total (lab)",
  name = "location_id", date = "sample_date")
}
\arguments{
\item{df}{dataframe of major ions gathered using \code{\link{get_major_ions}}}

\item{Mg}{magnesium}

\item{Ca}{calcium}

\item{K}{potassium}

\item{Cl}{chloride}

\item{SO4}{sulfate}

\item{HCO3}{alkalininity or bicarbonate}

\item{name}{column of sample location id}

\item{date}{column of sample date}
}
\description{
Function to transoform data from \code{\link{get_major_ions}} to schoeller 
coordinates
}
