% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{tukey_outlier}
\alias{tukey_outlier}
\title{Tukey's test for outliers}
\usage{
tukey_outlier(
  df,
  x = "analysis_result",
  k = 3,
  na.rm = FALSE,
  type = 6,
  combine_locations = FALSE
)
}
\arguments{
\item{df}{data.frame of groundwater data}

\item{x}{column of analysis results}

\item{k}{multiplier for IQR k = 1.5 indicates "outlier", k = 3 indicates
"far out"}

\item{na.rm}{logical. Should missing values be removed?}

\item{type}{an integer selecting one of the many quantile algorithms.
Default is 6 to match Sanitas}

\item{combine_locations}{TRUE/FALSE}
}
\description{
Tukey's test for outliers
}
