% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_background.R
\name{update_background}
\alias{update_background}
\title{Function to update background data}
\usage{
update_background(df, well, param, start = NULL, n_end = 8, n_prop = 4)
}
\arguments{
\item{df}{data frame of groundwater data in tidy format}

\item{well}{the location to update}

\item{param}{the constituent to update}

\item{start}{the start date of the background data. IF not provide the 
start of the data will be found using \code{\link{start_date()}}}

\item{n_end}{the number of samples in current background. The date will be 
found from this sample.}

\item{n_prop}{the number of samples to include in the proposed background}
}
\description{
Function to update background data
}
