% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{write_event_summary}
\alias{write_event_summary}
\title{Function to export summary table for a sampling event}
\usage{
write_event_summary(df, start, end, gw_elev = TRUE, short_name = TRUE, file)
}
\arguments{
\item{df}{groundwater data frame in tidy format}

\item{start}{start date}

\item{end}{end date}

\item{gw_elev}{if TRUE, list date well sampled and annotate date for GW Elev}

\item{short_name}{TRUE/FALSE to abbreviate constituent name}
}
\description{
Function to export summary table for a sampling event
}
