% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{write_excel}
\alias{write_excel}
\title{Function to export data from manages to excel. Each sampling location 
is saved to a tab.}
\usage{
write_excel(df, wells, constituents, file, short_name = TRUE)
}
\arguments{
\item{df}{groundwater data frame in tidy format}

\item{wells}{list of wells to be exported}

\item{constituents}{list of constituents to be exported}

\item{file}{full file path name with extension for export}

\item{short_name}{TRUE/FALSE to abbreviate constituent name}
}
\description{
Function to export data from manages to excel. Each sampling location 
is saved to a tab.
}
