% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_limits.R
\name{assign_limits}
\alias{assign_limits}
\title{Function to assign default Groundwater Protection Standards}
\usage{
assign_limits(
  df,
  column = "param_name",
  method = c("default", "tolerance"),
  locations = NULL
)
}
\arguments{
\item{df}{dataframe of groundwater data in tidy format}

\item{column}{column to search for constituents}

\item{method}{used to assign GWPS. Options c("default", "tolerance")}

\item{locations}{locations to be used}
}
\description{
Function to assign default Groundwater Protection Standards
}
