% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox_transform.R
\name{boxcox_transform}
\alias{boxcox_transform}
\title{Boxcox transformation}
\usage{
boxcox_transform(df, combine_locations = FALSE,
  keep_data_object = FALSE)
}
\arguments{
\item{df}{data frame}

\item{keep_data_object}{}
}
\description{
Boxcox transformation
}
