% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{boxplot}
\alias{boxplot}
\title{Function to plot boxplots for groundwater data}
\usage{
boxplot(df, x = "location_id", y = "analysis_result", fill = NULL,
  short_name = FALSE, coord_flip = FALSE, legend_title = NULL)
}
\arguments{
\item{df}{groundwater data in tidy, long format}

\item{x}{column to be used for the x axis}

\item{y}{column to be used for the y axis}

\item{fill}{column used to fill the variable}

\item{short_name}{If TRUE the constituent name will be abbreviated}

\item{legend_title}{Legend title for fill variable}

\item{flip_coords}{If TRUE the axes are flipped}
}
\description{
Function to plot boxplots for groundwater data
}
