% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{boxplot}
\alias{boxplot}
\title{Function to plot multiple boxplots for groundwater data}
\usage{
boxplot(df, facet_by = "location_id", short_name = FALSE,
  coord_flip = FALSE)
}
\arguments{
\item{df}{groundwater data}

\item{facet_by}{parameter to group data by}

\item{name}{If name = "short" the analyte will be abbreviated}

\item{flip_coords}{If TRUE the axes are flipped}
}
\description{
Function to plot multiple boxplots for groundwater data
}

