% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mcl.R
\name{compare_mcl}
\alias{compare_mcl}
\title{Function to compare result to EPA's MCL}
\usage{
compare_mcl(df, format = "summary", type = "all")
}
\arguments{
\item{df}{dataframe of groundwater data that has been unit checked}

\item{format}{default summary}

\item{type}{default is all}
}
\description{
Function to compare result to EPA's MCL
}

