% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_to_meq.R
\name{conc_to_meq}
\alias{conc_to_meq}
\title{Function to convert chemical data from mg/L to meq/L}
\usage{
conc_to_meq(df, Mg = "Magnesium, dissolved", Ca = "Calcium, dissolved",
  Na = "Sodium, dissolved", K = "Potassium, dissolved",
  Cl = "Chloride, total", SO4 = "Sulfate, total",
  Alk = "Alkalinity, total (lab)")
}
\arguments{
\item{df}{data frame}

\item{Mg}{Magnesium}

\item{Ca}{Calcium}

\item{Na}{Sodium}

\item{K}{Potassium}

\item{Cl}{Chloride}

\item{SO4}{Sulfate}

\item{Alk}{Total Alkalinity}
}
\description{
Function to convert chemical data from mg/L to meq/L
}
