% Generated by roxygen2 (4.0.2): do not edit by hand
\name{conc_to_meq}
\alias{conc_to_meq}
\title{function to convert geochemical plot data into meq/L}
\usage{
conc_to_meq(df, Mg = "Magnesium, dissolved", Ca = "Calcium, dissolved",
  Na = "Sodium, dissolved", K = "Potassium, dissolved",
  Cl = "Chloride, total", SO4 = "Sulfate, total",
  HCO3 = "Alkalinity, total (lab)")
}
\arguments{
\item{df}{data frame}

\item{Mg}{Magnesium}

\item{Ca}{Calcium}

\item{Na}{Sodium}

\item{K}{Potassium}

\item{Cl}{Chloride}

\item{SO4}{Sulfate}

\item{HCO3}{Bicaronate}
}
\description{
function to convert geochemical plot data into meq/L
}

