% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{conf_int}
\alias{conf_int}
\title{Function to calculate confidence intervals}
\usage{
conf_int(
  df,
  analysis_result = "analysis_result",
  distribution = "distribution",
  method = "mvue",
  ci_type = "two-sided",
  conf_level = 0.95
)
}
\arguments{
\item{df}{df data frame of groundwater data in tidy format}

\item{analysis_result}{the analysis result column}

\item{distribution}{the distribution column}

\item{method}{character string specifying the method of estimation.
Possible values are "mvue" (minimum variance unbiased; the default),
and "mle/mme" (maximum likelihood/method of moments).}

\item{ci_type}{character string indicating what kind of confidence interval
to compute. The possible values are ci_type="two-sided" (the default),
ci_type="lower", and ci_type = "upper}

\item{conf_level}{a scalar between 0 and 1 indicating the confidence level
of the prediction interval. The default value is conf_level = 0.95}
}
\description{
Function to calculate confidence intervals
}
\examples{
data("gw_data")
wells <- c("MW-1", "MW-2", "MW-3", "MW-4")
params <- c("Sulfate, total",
            "Arsenic, dissolved",
            "Boron, dissolved")

background <- lubridate::ymd(c("2007-12-20", "2012-01-01"), tz = "UTC")
# first group data by location, param, and background
# estimate percent less than and distribution
background_data <- gw_data \%>\%
 filter(location_id \%in\% wells, param_name \%in\% params,
         sample_date >= background[1] & sample_date <= background[2]) \%>\%
  group_by(location_id, param_name, default_unit) \%>\%
  percent_lt() \%>\%
  est_dist(., keep_data_object = TRUE) \%>\%
  arrange(location_id, param_name)

background_data \%>\%
conf_int(., ci_type = "lower", conf_level = 0.99)

}
