% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{delete_datebase}
\alias{delete_datebase}
\title{Function to delete an existing water quality database}
\usage{
delete_datebase(path)
}
\arguments{
\item{path}{path to sqlite database}
}
\description{
Function to delete an existing water quality database
}
