% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{dist}
\alias{dist}
\title{Function to return distribution based on using p-value}
\usage{
dist(x, p = 0.01)
}
\arguments{
\item{x}{x column of analysis result for groundwater data in tidy format}

\item{p}{p p-value for gofTest}
}
\description{
Function to return distribution based on using p-value
}
