% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality_censored.R
\name{dist_censored}
\alias{dist_censored}
\title{Helpre function to return distribution based on using p-value}
\usage{
dist_censored(x, censored, p = 0.01, ...)
}
\description{
Helpre function to return distribution based on using p-value
}
