% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality_censored.R
\name{dist_censored}
\alias{dist_censored}
\title{Function to return distribution based on using p-value}
\usage{
dist_censored(x, censored)
}
\arguments{
\item{x}{analysis_result of groundwater data}
}
\description{
Function to return distribution based on using p-value
}
