% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dist_censored}
\alias{dist_censored}
\title{Function to return distribution based on using p-value}
\usage{
dist_censored(x)
}
\arguments{
\item{x}{analysis_result of groundwater data}
}
\description{
Function to return distribution based on using p-value
}

