% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{est_dist}
\alias{est_dist}
\title{Function to return either normal, lognormal, or non-parametric multiple 
groundwater data locations and parameters.}
\usage{
est_dist(df, p = 0.01)
}
\arguments{
\item{df}{groundwater data frame}
}
\description{
Function to return either normal, lognormal, or non-parametric multiple 
groundwater data locations and parameters.
}

