% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_dist.R
\name{est_dist}
\alias{est_dist}
\title{Function to return either normal, lognormal, or non-parametric 
distribution.}
\usage{
est_dist(df, alpha = 0.05, method = "sw", choices = c("norm",
  "lnorm"), group_by_location = FALSE, keep_data_object = FALSE)
}
\arguments{
\item{df}{groundwater data frame in tidy format}

\item{alpha}{alpha}

\item{method}{default is "sw"}

\item{choices}{vector of distributions to check. Default is c("norm, "lnorm")}

\item{group_by_location}{TRUE/FALSE to estimate distribution by individual
location, or grouped together. Default is FALSE.}

\item{keep_data_object}{Default is FALSE}
}
\description{
Function to return either normal, lognormal, or non-parametric 
distribution.
}
\examples{
data("gw_data")
wells <- c("MW-1", "MW-2", "MW-3", "MW-4")
params <- c("Sulfate, total",
            "Arsenic, dissolved",
           "Boron, dissolved")

gw_data <- gw_data \%>\%
filter(location_id \%in\% wells, param_name \%in\% params) \%>\%
percent_lt() \%>\%
filter(percent_lt <=15)

gw_data \%>\%
group_by(location_id, param_name, default_unit) \%>\%
est_dist(., alpha = 0.05, keep_data_object = FALSE)


gw_data \%>\%
group_by(param_name, default_unit) \%>\%
est_dist(., group_by_location = TRUE, keep_data_object = FALSE)
}
