% Generated by roxygen2 (4.0.2): do not edit by hand
\name{est_dist_censored}
\alias{est_dist_censored}
\title{Function to return either normal, lognormal, or non-parametric multiple
groundwater data locations and parameters.}
\usage{
est_dist_censored(df)
}
\arguments{
\item{df}{groundwater data frame}
}
\description{
Function to return either normal, lognormal, or non-parametric multiple
groundwater data locations and parameters.
}

