% Generated by roxygen2 (4.0.1): do not edit by hand
\name{event_summary}
\alias{event_summary}
\title{Function to export summary table for a sampling event}
\usage{
event_summary(df, wells, constituents, start, end)
}
\arguments{
\item{df}{data frame in the format location_id, sample_date, param_name,
lt_measure, default_unit}

\item{wells}{list of wells}

\item{constituents}{list of constituents}

\item{start}{start date}

\item{end}{end date}
}
\description{
Function to export summary table for a sampling event
}

