% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{event_summary}
\alias{event_summary}
\title{Function to export summary table for a sampling event}
\usage{
event_summary(df, wells, params, start, end, gw_elev = TRUE)
}
\arguments{
\item{df}{groundwater data frame in tidy format}

\item{wells}{list of wells to include}

\item{params}{list of parameters to include}

\item{start}{start date}

\item{end}{end date}

\item{gw_elev}{if TRUE, list date well sampled and annotate date for GW Elev}
}
\description{
Function to export summary table for a sampling event
}
