% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{event_summary}
\alias{event_summary}
\title{Function to export summary table for a sampling event}
\usage{
event_summary(df, wells, constituents, start, end)
}
\arguments{
\item{df}{data frame in the format location_id, sample_date, param_name,
lt_measure, default_unit
@param wells list of wells
@param constituents list of constituents
@param start start date
@param end end date}
}
\description{
Function to export summary table for a sampling event
}

