\title{
Documentation source code of a function
}
\name{get_function}
\alias{get_function}
\description{
Get the source code of the documentation file (\code{.Rd}) of a given function as a list.
}\usage{
  get_function(func, pkg = NULL)
}\arguments{
  \item{func}{
    Function name as a character string.
}
}\value{
  A list
}\author{
  Hugo Marthinet
}\examples{
  get_function("get_function")
}
