% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochemistry_plots.R
\name{get_major_ions}
\alias{get_major_ions}
\title{Function to gather parameters needed for geochemistry plots.}
\usage{
get_major_ions(df, param_name, Mg = "Magnesium, dissolved",
  Ca = "Calcium, dissolved", Na = "Sodium, dissolved",
  K = "Potassium, dissolved", Cl = "Chloride, total",
  SO4 = "Sulfate, total", Alk = "Alkalinity, total (lab)", TDS = NULL)
}
\arguments{
\item{df}{data frame of groundwater monitoring data in tidy format}

\item{param_name}{column of constituents}

\item{Mg}{Magnesium}

\item{Ca}{Calcium}

\item{Na}{Sodium}

\item{K}{Potassium}

\item{Cl}{Chloride}

\item{SO4}{Sulfate}

\item{Alk}{Alkalinity}

\item{TDS}{Total Dissolved Solids}
}
\description{
Function to gather parameters needed for geochemistry plots.
}
