% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_major_ions}
\alias{get_major_ions}
\title{function to read monitoring well data in the form location_id, sample_date,
analysis_result, default_unit, param_name and gather parameters needed for
geochemistry plots.
The data is cast by location_id + sample_date ~ param_name}
\usage{
get_major_ions(df, Mg = "Magnesium, dissolved", Ca = "Calcium, dissolved",
  Na = "Sodium, dissolved", K = "Potassium, dissolved",
  Cl = "Chloride, total", SO4 = "Sulfate, total",
  Alk = "Alkalinity, total (lab)", TDS = "Total Dissolved Solids", ...)
}
\arguments{
\item{df}{data frame of groundwater monitoring data}

\item{Mg}{Magnesium}

\item{Ca}{Calcium}

\item{Na}{Sodium}

\item{K}{Potassium}

\item{Cl}{Chloride}

\item{SO4}{Sulfate}

\item{Alk}{Alkalinity}

\item{TDS}{Total Dissolved Solids}
}
\description{
function to read monitoring well data in the form location_id, sample_date,
analysis_result, default_unit, param_name and gather parameters needed for
geochemistry plots.
The data is cast by location_id + sample_date ~ param_name
}

