% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_mcl_params}
\alias{get_mcl_params}
\title{Function to find EPA primary and seconday water constituents}
\usage{
get_mcl_params(df, type = "all")
}
\arguments{
\item{df}{dataframe of groundwater data}
}
\description{
Function to find EPA primary and seconday water constituents
}

