\title{
  Chemins vers le script actif
}\name{get_path}\alias{get_script_path}\alias{get_script_file_path}\alias{get_file_path}\description{
  \code{get_script_path()} retourne le chemin complet du script actif. \cr
  \code{get_script_file_path()} retourne le chemin complet du dossier du script actif. \cr
  \code{get_file_path()} retourne le chemin du dossier du script actif.
}\usage{
  get_script_path()
  get_script_file_path()
  get_file_path()
}\value{
  Retourne un chemin sous forme de chaîne de caractère.
}\author{
  Hugo Marthinet
}\examples{
## Aucun argument nécessaire
get_script_path()
get_script_file_path()
get_file_path()
}
