% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{get_theilsen}
\alias{get_theilsen}
\title{Function to get slope and intercept from Theil-Sen slope for plotting}
\usage{
get_theilsen(df, x = "ANALYSIS_RESULT", y = "SAMPLE_DATE", ...)
}
\arguments{
\item{df}{dataframe of groundwater data in tidy format}

\item{x}{column for analysis result}

\item{y}{column for sample date}

\item{...}{other arguements passed to \code{\link{kendallTrendTest}}}
}
\description{
Function to get slope and intercept from Theil-Sen slope for plotting
}
