% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{get_theilsen}
\alias{get_theilsen}
\title{Function to get slope and intercept from Theil-Sen slope for plotting}
\usage{
get_theilsen(df, x = "analysis_result", y = "sample_date", ...)
}
\description{
Function to get slope and intercept from Theil-Sen slope for plotting
}
