% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{get_wells}
\alias{get_wells}
\title{Function to return a list of all the location IDs}
\usage{
get_wells(df)
}
\arguments{
\item{df}{data frame of groundwater data in the format with column names
location_id, param_name, default_unit, lt_measure, analysis_result}
}
\description{
Function to return a list of all the location IDs
}

