% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_inflix.R
\name{\%+\%}
\alias{\%+\%}
\title{\%+\%}
\usage{
... \%+\% NA
}
\arguments{
\item{...}{Strings (or string-coercible objects) placed on either side of the \%+\% operator.}
}
\value{
A character vector with length equal to the maximum length of the inputs.
}
\description{
String-concatenation inflix operator a la '+' in Python.
}
\details{
No dependencies.
}
\examples{
\dontrun{
'aaa' \%+\% 'bbb' \%+\% 'ccc' \%+\% 'ddd'
# Zero-length inputs are concatenated in the resulting string as ''
'aaa' \%+\% 'bbb' \%+\% 'ccc' \%+\% 'ddd'[F]
# If an input has length > 1, the result has length > 1 and a warning is given
c('aaa', 'bbb') \%+\% 'zzz'
}

}
