% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gw_ts_plot}
\alias{gw_ts_plot}
\title{Function for plotting time series of groundwater data}
\usage{
gw_ts_plot(df, facet_by = NULL, back_date = NULL, comp_date = NULL,
  limit1 = NULL, limit2 = NULL, short_name = FALSE, pnt = 3,
  ncol = NULL)
}
\arguments{
\item{df}{groundwater data}

\item{facet_by}{parameter to group plots by}

\item{back_date}{dates for background date range}

\item{comp_date}{dates for compliance date range}

\item{limit1}{horizontal line 1}

\item{limit2}{horizontal line 2}

\item{short_name}{If TRUE, the analyte name will be abbreviated}

\item{pnt}{size of points on time series plots}

\item{ncol}{number of columns}
}
\description{
Function for plotting time series of groundwater data
}

