% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{include_dup}
\alias{include_dup}
\title{Function to include duplicate samples}
\usage{
include_dup(df, wells)
}
\arguments{
\item{df}{groundwater data frame}

\item{wells}{list of wells to included duplicates for}
}
\description{
Function to include duplicate samples
}
