% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indiana_data}
\alias{indiana_data}
\title{Groundwater quality data for the state of Indiana}
\format{A data frame with 162923 rows and 11 variables:
\describe{
  \item{SiteId}{Site Identifier}
  \item{Result}{analytical result}
  ...
}}
\source{
\url{http://in.gov/idem/cleanwater/2453.html}
}
\usage{
indiana_data
}
\description{
A dataset containing groundwater quality data for the state of Indiana.
}
\keyword{datasets}
