% Generated by roxygen2 (4.0.2): do not edit by hand
\name{intra_pred_int}
\alias{intra_pred_int}
\title{Function to calculate intrawell prediction interval}
\usage{
intra_pred_int(df, analysis_result, wells, params, bkgd_dates, comp_dates,
  non_detect = c(15, 50), cen_method = "mle", intra.conf.level = 0.95,
  simultaneous = TRUE, SWFPR = NULL, ...)
}
\arguments{
\item{df}{data frame of groundwater data}

\item{analysis_result}{name of column containing analysis results}

\item{wells}{vector of wells to be included}

\item{params}{vector of constituents to be included}

\item{bkgd_dates}{background data date range}

\item{comp_dates}{compliance data date range}

\item{intra.conf.level}{confidence level}

\item{SWFPR}{Site-Wide False Positive Rate. If supplied, then
intra.conf.level is computed using (1 - SWFPR)^(1/(nc*nw)) where nc is the
number of constituents and nw is the number of wells}
}
\description{
Function to calculate intrawell prediction interval
}

