% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{is_lognormal}
\alias{is_lognormal}
\title{Function to check if groundwater data has a Lognormal distribution}
\usage{
is_lognormal(x, p = 0.01)
}
\arguments{
\item{x}{analysis_result of groundwater data}
}
\description{
Function to check if groundwater data has a Lognormal distribution
}
