% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{is_lognormal}
\alias{is_lognormal}
\title{Helper function to check if data has a Lognormal distribution}
\usage{
is_lognormal(x, p = 0.01)
}
\description{
Helper function to check if data has a Lognormal distribution
}
