% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality_censored.R
\name{is_lognormal_censored}
\alias{is_lognormal_censored}
\title{Function to check if censored groundwater data has a Lognormal distribution}
\usage{
is_lognormal_censored(x, censored, ...)
}
\arguments{
\item{x}{analysis_result of groundwater data}

\item{censored}{logical vector of censored values}
}
\description{
Function to check if censored groundwater data has a Lognormal distribution
}
