% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{is_normal}
\alias{is_normal}
\title{Function to check if groundwater data has a Normal distribution}
\usage{
is_normal(x, p = 0.01)
}
\arguments{
\item{x}{x column of analysis result for groundwater data in tidy format}

\item{p}{p p-value for gofTest}
}
\description{
Function to check if groundwater data has a Normal distribution
}
