% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is_normal}
\alias{is_normal}
\title{Function to check if groundwater data has a Normal distribution}
\usage{
is_normal(x)
}
\arguments{
\item{x}{analysis_result of groundwater data}
}
\description{
Function to check if groundwater data has a Normal distribution
}

