% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality_censored.R
\name{is_normal_censored}
\alias{is_normal_censored}
\title{Helper function to check if censored groundwater data has a
Normal distribution}
\usage{
is_normal_censored(x, censored, p = 0.01, ...)
}
\description{
Helper function to check if censored groundwater data has a
Normal distribution
}
