% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{join_lt}
\alias{join_lt}
\title{Function to join columns of LT_MEASURE and sample results}
\usage{
join_lt(df, col_name)
}
\arguments{
\item{df}{groundwater data frame}

\item{col_name}{qouted column name for the result}
}
\description{
Function to join columns of LT_MEASURE and sample results
}
