% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loch_missingness_monster.R
\name{loch_missingness_monster}
\alias{loch_missingness_monster}
\title{loch_missingness_monster}
\usage{
loch_missingness_monster(dat, percent = NA)
}
\arguments{
\item{dat}{A data frame or matrix.}

\item{percent}{A number (optional).}
}
\value{
\code{loch_missingness_monster()} just returns output in the console.
}
\description{
Provides an easy-to-interpret in-console breakdown of missingness in data sets.
}
\details{
\code{loch_missingness_monster()} tells you:
\itemize{
  \item The total number of missing values (\code{NA}) in the data set.
  \item A table indicating the count and percent of missing values per column.
  \item Optionally, the number of rows missing values in excess of a percent that you specify.
}
(The output comes in sentence form if there are less or equal to than 10 variables; it comes in table form if there are more than 10.)
}
\examples{
lmm_example <- replicate(12, sample(1:10, 3, replace = TRUE))
lmm_example[sample(length(lmm_example), 13, replace = FALSE)] <- NA
colnames(lmm_example) <- paste0('var', 1:12, sep = '')
loch_missingness_monster(lmm_example, 40)

}
