% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loch_missingness_monster.R
\name{loch_missingness_monster}
\alias{loch_missingness_monster}
\title{loch_missingness_monster}
\usage{
loch_missingness_monster(data, percent)
}
\arguments{
\item{data}{A data frame.}

\item{percent}{A percent (numeric).}
}
\value{
\code{loch_missingness_monster()} just returns output in the console.
}
\description{
Provides an easy-to-interpret in-console breakdown of missingness in data sets.
}
\details{
\code{loch_missingness_monster()} will tell you:
\itemize{
  \item The total number of missing values (NA) in the data set.
  \item The maximum number of values that any variable is missing.
  \item How many values each variable is missing.
  \item Which rows are missing values in excess of a percent that you specify.
}
(The output comes in sentence form if there are less or equal to than 10 variables, and table form if there are more than 10.)
}
\examples{
# Case with <=10 variables:
lmm_example <- data.frame(c(1, NA, 3, 3, 6, 3),
                            c(3, NA, NA, NA, NA, 6),
                            c(3, 7, 2, NA, NA, 6))
colnames(lmm_example) <- c("var1", "var2", "var3")
loch_missingness_monster(lmm_example, 65)

# Case with > 10 variables:
set.seed <- 4
lmm_example2 <- replicate(12, sample(1:10, 3, replace = TRUE))
lmm_example2[sample(length(lmm_example2), 13, replace = FALSE)] <- NA
colnames(lmm_example2) <- paste0('var', 1:12, sep = '')
loch_missingness_monster(lmm_example2, 40)

}
