\name{manageR-package}

\alias{manageR-package}
\alias{manageR}

\docType{package}

\title{
  manageR
}

\description{
  Ensemble de fonctions permettant une meilleure gestion des fichiers, des paramètres de R et des libraries.
}

\details{
  \tabular{ll}{
    Package: \tab  manageR \cr
    Type: \tab
    Package\cr
    Version: \tab  0.1.0 \cr
    Date: \tab  01/10/2022 \cr
    License: \tab neant \cr
    LazyLoad: \tab yes\cr
  }


}

\author{
  Hugo Marthinet  <h.marthinet@gmail.com> \cr
}

\references{

}

\seealso{
  Autres packages developpes au CESPA :

  \code{\link{epiplus}}, \code{\link{MDBprod}}, \code{\link{lazyMe}},
  \code{\link{surveyscores}}, \code{\link{NSN}}
}

\examples{

}
