\name{manageR-package}

\alias{manageR-package}
\alias{manageR}

\docType{package}

\title{
  manageR
}

\description{
  File, path, env and package handling.
}

\details{
  \tabular{ll}{
    Package: \tab  manageR \cr
    Type: \tab
    Package\cr
    Version: \tab  0.2.0 \cr
    Date: \tab  01/10/2022 \cr
    License: \tab neant \cr
    LazyLoad: \tab yes\cr
  }


}

\author{
  Hugo Marthinet  <h.marthinet@gmail.com> \cr
}

\references{

}

\seealso{
  Other packages by the same author :

  \code{\link{epiplus}}, \code{\link{dataDesc}}, \code{\link{lazyMe}},
  \code{\link{surveyscores}}, \code{\link{sampleSize}}
}

\examples{

}
