% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager-package.R, R/manager_shiny_app.R
\docType{package}
\name{manager}
\alias{manager}
\alias{manager-package}
\title{manager}
\usage{
manager()
}
\description{
Tools for plotting and analyzing environmental
    data as well as for reading data from external sources such as 
    MANAGES and gINT.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/jentjr/manager}
}

}
\author{
\strong{Maintainer}: Justin Jent \email{jentjr@gmail.com}

}
