\name{mcl_heatmap}
\alias{mcl_heatmap}
\title{Plot for each mcl with location on x axis, MCL parameters on x axis}
\usage{
mcl_heatmap(df, title)
}
\arguments{
  \item{df}{dataframe of mcl summary}

  \item{title}{title of plot}
}
\description{
Plot for each mcl with location on x axis, MCL parameters
on x axis
}

