% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_edit_settingsInput.R
\name{mod_edit_settingsInput}
\alias{mod_edit_settingsInput}
\alias{mod_edit_settings}
\title{mod_edit_settingsInput and mod_edit_settings}
\usage{
mod_edit_settingsInput(id)

mod_edit_settings(input, output, session, settingsdata, package)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{settingsdata}{the reactive xml settings file}

\item{package}{package where the xml config file is}
}
\value{
the modified reactive xml settings file
}
\description{
A shiny module to manage a settings file stored as xml in the www
folder of the package
}
\examples{
\dontrun{
library(shiny)
library(xml2)
if (interactive()) {
ui <- fluidPage(
  mod_edit_settingsInput("config")
)

server <- function(input, output, session) {
  initsetsdata <- reactive(read_xml(system.file("www/config.xml", package = "manageR")))
  setsdata <- callModule(mod_edit_settings, 
                         "config", 
                         settingsdata = initsetsdata,
                         package = "manageR")
}

shinyApp(ui, server)
}
}
}
