% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{name_units}
\alias{name_units}
\title{Function to return a column of parameter name with units}
\usage{
name_units(df, short_name = TRUE)
}
\arguments{
\item{df}{groundwater data frame}

\item{short_name}{If TRUE will use abbreviated constituent names}
}
\description{
Function to return a column of parameter name with units
}
