% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_person_regex.R
\docType{data}
\name{non_person_regex}
\alias{non_person_regex}
\title{Regex pattern for identifying non-person names}
\format{
A character string
}
\source{
Created by github.com/jacob-gg
}
\usage{
non_person_regex
}
\description{
A regex pattern that matches names that are likely to be non-persons (e.g., businesses, educational institutions, government entities, etc.)
}
\details{
User \code{check_regex()} to check \code{non_person_regex} against a test data set, \code{test_names}.
}
\keyword{datasets}
