% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ohio_data}
\alias{ohio_data}
\title{Groundwater quality data for the state of Ohio}
\format{A data frame with 213845 rows and 47 variables:
\describe{
  \item{well_id}{well_id, Site Identifier}
  \item{result_value}{analytical result}
  ...
}}
\source{
\url{http://epa.ohio.gov/ddagw/gwqcp.aspx}
}
\usage{
ohio_data
}
\description{
A dataset containing groundwater quality data for the state of Ohio.
}
\keyword{datasets}
