% Generated by roxygen2 (4.0.2): do not edit by hand
\name{percent_gt}
\alias{percent_gt}
\title{calculate the percentage of right censored data}
\usage{
percent_gt(gt)
}
\arguments{
\item{gt}{column of data greater than detection limit.}

\item{df}{data frame of groundwater monitoring data in long format}
}
\description{
calculate the percentage of right censored data
}

