% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{percent_gt}
\alias{percent_gt}
\title{Calculate the percentage of right censored data}
\usage{
percent_gt(df, lt_measure = "lt_measure", percent_gt = "percent_gt")
}
\arguments{
\item{df}{df data frame of groundwater monitoring data in long format}

\item{lt_measure}{lt_measure column of greater than detection limit symbol.}

\item{percent_gt}{percent_gt new column name}
}
\description{
Calculate the percentage of right censored data
}
