\name{percent_lt}
\alias{percent_lt}
\title{calculate the percentage of non-detects}
\usage{
percent_lt(lt)
}
\arguments{
  \item{df}{data frame of groundwater monitoring data in
  long format}

  \item{lt}{column of non-detects. Default is set to
  lt_measure which has the "<" symbol.}
}
\description{
calculate the percentage of non-detects
}

