% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{percent_lt}
\alias{percent_lt}
\title{Calculate the percentage of left censored data}
\usage{
percent_lt(df, lt_measure)
}
\arguments{
\item{df}{df data frame of groundwater monitoring data in long format}

\item{lt_measure}{lt_measure column of less than detection limit symbol.}
}
\description{
Calculate the percentage of left censored data
}
