% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper_plot.R
\name{piper_time_plot}
\alias{piper_time_plot}
\title{Function to create an animated Piper plot through time
using the animation package}
\usage{
piper_time_plot(df, ...)
}
\arguments{
\item{df}{data frame of groundwater data transformed using}

\item{...}{
  Arguments passed on to \code{\link[=piper_plot]{piper_plot}}
  \describe{
    \item{\code{location_id}}{column for sample location}
    \item{\code{sample_date}}{column for sample date}
    \item{\code{x_cation}}{default is Calcium, dissolved}
    \item{\code{x_cation_label}}{label for x cation}
    \item{\code{y_cation}}{default is Magnesium, dissolved}
    \item{\code{y_cation_label}}{label for y cation}
    \item{\code{z_cation}}{default is Sodium, dissolved + Potassium, dissolved}
    \item{\code{z_cation_label}}{label for z cation}
    \item{\code{x_anion}}{default is Chloride, total + Fluoride, total}
    \item{\code{x_anion_label}}{label for x anion}
    \item{\code{y_anion}}{default is Alkalinity, total (lab)}
    \item{\code{y_anion_label}}{label for y anion}
    \item{\code{z_anion}}{default is Sulfate, total}
    \item{\code{z_anion_label}}{label for z anion}
    \item{\code{x_z_anion_label}}{label for the upper left diamond}
    \item{\code{x_y_cation_label}}{label for the upper right diamond}
    \item{\code{total_dissolved_solids}}{Scale plot by Total Dissolved Solids,
default = FALSE}
    \item{\code{group_col}}{grouping column for symbols}
    \item{\code{transparency}}{the setting for transparency value for points. Default
is 0.2}
    \item{\code{pnt_size}}{the size of the points. Default is 3}
    \item{\code{label_size}}{size of font for labels}
    \item{\code{title}}{Title for plot, default = NULL}
  }}
}
\description{
Function to create an animated Piper plot through time
using the animation package
}
