\name{piper_time_plot}
\alias{piper_time_plot}
\title{Function to created an animated Piper plot through time
using the animation package}
\usage{
piper_time_plot(df, TDS = FALSE)
}
\arguments{
  \item{df}{data frame of groundwater data transformed
  using \code{\link{transform_piper_data}}}
}
\description{
Function to created an animated Piper plot through time
using the animation package
}

