% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochemistry_plots.R
\name{piper_time_plot}
\alias{piper_time_plot}
\title{Function to create an animated Piper plot through time
using the animation package}
\usage{
piper_time_plot(df, TDS = FALSE, title = NULL)
}
\arguments{
\item{df}{data frame of groundwater data transformed using 
\code{\link{transform_piper_data}}}

\item{TDS}{Scale by Total Dissolved Solids}
}
\description{
Function to create an animated Piper plot through time
using the animation package
}
