% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pred_int}
\alias{pred_int}
\title{Function to calculate prediction interval}
\usage{
pred_int(x, n.mean = 1, k = 1, m = 1, method = "Bonferroni",
  pi.type = "upper", conf.level = 0.95, dist = NULL)
}
\arguments{
\item{x}{analysis_result of gw data.frame}

\item{n.mean}{positive integer specifying the sample size associated
with the future averages.
The default value is n.mean=1 (i.e., individual observations).
Note that all future averages must be based on the same sample size.}

\item{k}{positive integer specifying the number of future observations or
averages the prediction interval should contain with confidence level
conf.level. The default value is k=1.}

\item{pi.type}{character string indicating what kind of prediction interval
to compute. The possible values are pi.type="upper" (the default),
and pi.type="lower".}

\item{conf.level}{a scalar between 0 and 1 indicating the confidence level
of the prediction interval. The default value is conf.level=0.95}

\item{dist}{override distribution. Default is NULL, which will check for
distribution.}
}
\description{
Function to calculate prediction interval
}

