% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_int.R
\name{pred_int_sim}
\alias{pred_int_sim}
\title{Function to calculate simultaneous prediction interval}
\usage{
pred_int_sim(x, left_censored, percent_left, non_detect = c(15, 50),
  cen_method = "mle", dist = NULL, conf.level = 0.95, ...)
}
\arguments{
\item{x}{analysis_result of gw data.frame}

\item{left_censored}{column of TRUE/FALSE for censored variables}

\item{percent_left}{column of percent left censored variables}

\item{non_detect}{vector of cut-off values for non-detect processing}

\item{cen_method}{method for handlin non-detected between the cut-off values}

\item{...}{other variabled passed to EnvStats predInt... functions}
}
\description{
Function to calculate simultaneous prediction interval
}
