% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pred_int_sim}
\alias{pred_int_sim}
\title{Function to calculate simultaneous prediction interval}
\usage{
pred_int_sim(x, left_censored, non_detect = c(15, 50), cen_method = "mle",
  n.mean = 1, k = 1, m = 2, r = 2, rule = "k.of.m",
  pi.type = "upper", conf.level = 0.95, dist = NULL,
  K.tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{analysis_result of gw data.frame}

\item{n.mean}{positive integer specifying the sample size associated
with the future averages.
The default value is n.mean=1 (i.e., individual observations).
Note that all future averages must be based on the same sample size.}

\item{k}{for the k-of-m rule (rule="k.of.m"), a positive integer specifying
the minimum number of observations (or averages) out of m observations
(or averages) (all obtained on one future sampling “occassion”) the
prediction interval should contain with confidence level conf.level.
The default value is k=1. This argument is ignored when the argument rule
is not equal to "k.of.m".}

\item{m}{positive integer specifying the maximum number of future
observations (or averages) on one future sampling “occasion”.
The default value is m = 2, except when rule="Modified.CA", in which case
this argument is ignored and m is automatically set equal to 4.}

\item{r}{positive integer specifying the number of future sampling
“occasions”. The default value is r=2.}

\item{rule}{character string specifying which rule to use.
The possible values are "k.of.m" (k-of-m rule; the default),
"CA" (California rule), and "Modified.CA" (modified California rule).}

\item{pi.type}{character string indicating what kind of prediction interval
to compute. The possible values are pi.type="upper" (the default),
and pi.type="lower".}

\item{conf.level}{a scalar between 0 and 1 indicating the confidence level
of the prediction interval. The default value is conf.level=0.95}

\item{dist}{override distribution. Default is NULL, which will check for
distribution.}

\item{K.tol}{numeric scalar indicating the tolerance to use in the nonlinear
search algorithm to compute K. The default value is
K.tol=.Machine$double.eps^(1/2). For many applications, the value of K needs
to be known only to the second decimal place, in which case
setting K.tol=1e-4 will speed up computation a bit.}
}
\description{
Function to calculate simultaneous prediction interval
}

