\name{remove_dup}
\alias{remove_dup}
\title{function to remove duplicate samples
Example: If you have wells named MW-1 and another named MW-1 Duplicate
this function will remove the MW-1 Duplicate sample}
\usage{
remove_dup(df)
}
\arguments{
  \item{df}{data frame of groundwater data in long format
  with location_id as the column name for monitoring wells}
}
\description{
function to remove duplicate samples Example: If you have
wells named MW-1 and another named MW-1 Duplicate this
function will remove the MW-1 Duplicate sample
}

