% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{replace_missing}
\alias{replace_missing}
\title{Function to replace missing values}
\usage{
replace_missing(df)
}
\arguments{
\item{df}{groundwater data frame}
}
\description{
Function to replace missing values
}
