% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoeller_plot.R
\name{schoeller_plot}
\alias{schoeller_plot}
\title{Function to plot data transformed to schoeller coordinates using
\code{\link{.transform_schoeller_data}}}
\usage{
schoeller_plot(df, LOCATION_ID = "LOCATION_ID", SAMPLE_DATE = "SAMPLE_DATE",
  PARAM_NAME = "PARAM_NAME", ANALYSIS_RESULT = "ANALYSIS_RESULT",
  DEFAULT_UNIT = "DEFAULT_UNIT", Ca = "Calcium, dissolved",
  Mg = "Magnesium, dissolved", Na = "Sodium, dissolved",
  K = "Potassium, dissolved", Cl = "Chloride, total",
  Alk = "Alkalinity, total (lab)", SO4 = "Sulfate, total",
  facet_by = NULL, title = NULL, lwt = 1)
}
\arguments{
\item{df}{dataframe of groundwater data transformed to Schoeller coordinates}

\item{facet_by}{parameter to facet plots by}

\item{title}{title of plot}

\item{lwt}{lineweight}
}
\description{
Function to plot data transformed to schoeller coordinates using
\code{\link{.transform_schoeller_data}}
}
