% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoeller_plot.R
\name{schoeller_plot}
\alias{schoeller_plot}
\title{Function to plot data transformed to schoeller coordinates using
\code{\link{.transform_schoeller_data}}}
\usage{
schoeller_plot(df, location_id = "location_id",
  sample_date = "sample_date", param_name = "param_name",
  analysis_result = "analysis_result", default_unit = "default_unit",
  calcium = "Calcium, dissolved", magnesium = "Magnesium, dissolved",
  sodium = "Sodium, dissolved", potassium = "Potassium, dissolved",
  chloride = "Chloride, total", alkalinity = "Alkalinity, total (lab)",
  sulfate = "Sulfate, total", facet_var = "location_id",
  title = NULL, lwt = 1)
}
\arguments{
\item{df}{dataframe of environmental data in tidy format. The function
will convert data from mg/L to meq/L}

\item{location_id}{the column for sample locations}

\item{sample_date}{column for sample date}

\item{param_name}{the column of parameter names}

\item{analysis_result}{column containing the numerical analysis result}

\item{default_unit}{column containing the units. Assumes mg/L.}

\item{calcium}{name of calcium. Default is "Calcium, dissolved"}

\item{magnesium}{name of magnesium. Default is "Magnesium, dissolved"}

\item{sodium}{name for sodium. Default is "Sodium, dissolved"}

\item{potassium}{name for potassium. Default is "Potassium, dissolved"}

\item{chloride}{name for chloride. Default is "Chloride, total"}

\item{alkalinity}{name for alkalinity. Default is "Alkalinity, total (lab)"}

\item{sulfate}{name of sulfate. Default is "Sulfate, total"}

\item{facet_var}{parameter to facet plots by. Default is location}

\item{title}{title of plot}

\item{lwt}{lineweight}
}
\description{
Function to plot data transformed to schoeller coordinates using
\code{\link{.transform_schoeller_data}}
}
\examples{
data(gw_data)
gw_data \%>\%
filter(location_id \%in\% c("MW-1", "MW-2")) \%>\%
schoeller_plot(., facet_var = "location_id", title = "Example Scholler Plot")
 
gw_data \%>\%
filter(location_id \%in\% c("MW-1", "MW-2")) \%>\%
schoeller_plot(., facet_var = "sample_date", lwt = 2)

}
