% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochemistry_plots.R
\name{series_plot}
\alias{series_plot}
\title{Function to plot a series plot of groundwater data}
\usage{
series_plot(df, facet_by = NULL)
}
\arguments{
\item{df}{datadframe of groundwater data}

\item{facet_by}{parameter to facet the plot by}
}
\description{
Function to plot a series plot of groundwater data
}
