% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_helper.R
\name{set_background}
\alias{set_background}
\title{Function to return a column of background data}
\usage{
set_background(df, SAMPLE_DATE, start, end, return_all = FALSE)
}
\arguments{
\item{df}{dataframe of groundwater data in tidy format}

\item{SAMPLE_DATE}{the column for the sampling dates}

\item{start}{the start date for background}

\item{end}{the end date for background}

\item{return_all}{logical condition to return all values, or only background}
}
\description{
Function to return a column of background data
}
