% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censored_helper.R
\name{set_censored}
\alias{set_censored}
\title{Function to ensure non-detect data are set to either the PQL, or RL}
\usage{
set_censored(df, data_qualifier = "data_qualifier",
  analysis_result = "analysis_result", censor_column = "RL")
}
\arguments{
\item{df}{data frame of groundwater data in tidy format}

\item{data_qualifier}{column for data validation qualifier(i.e. U, or J Flag)}

\item{analysis_result}{column for analysis results}

\item{censor_column}{column of the censoring limit}
}
\description{
Function to ensure non-detect data are set to either the PQL, or RL
}
