% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_helper.R
\name{start_date}
\alias{start_date}
\title{Function to find the start of sampling for a well constituent pair}
\usage{
start_date(df, sample_date)
}
\arguments{
\item{df}{df dataframe of groundwater data in tidy format}

\item{sample_date}{sample_date the column for the sampling dates}
}
\description{
Function to find the start of sampling for a well constituent pair
}
