% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiff_plot.R
\name{stiff_plot}
\alias{stiff_plot}
\title{Stiff Diagram}
\usage{
stiff_plot(df, location_id = "location_id",
  sample_date = "sample_date", param_name = "param_name",
  analysis_result = "analysis_result", default_unit = "default_unit",
  magnesium = "Magnesium, dissolved", calcium = "Calcium, dissolved",
  sodium = "Sodium, dissolved", potassium = "Potassium, dissolved",
  chloride = "Chloride, total", sulfate = "Sulfate, total",
  alkalinity = "Alkalinity, total (lab)",
  total_dissolved_solids = NULL, group_var = "location_id",
  facet_var = "sample_date", lines = FALSE, cex = 1)
}
\arguments{
\item{df}{data frame of water quality data}

\item{location_id}{column for sample locations}

\item{sample_date}{column for sample date}

\item{param_name}{column for constituent names}

\item{analysis_result}{column for analysis result}

\item{default_unit}{column for units. Assumes all units are in mg/L and
converts usign \code{\link{conc_to_meq}}}

\item{magnesium}{name for magnesium}

\item{calcium}{name for calcium}

\item{sodium}{name for sodium}

\item{potassium}{name for potassium}

\item{chloride}{name for chloride}

\item{sulfate}{name for sulfate}

\item{alkalinity}{name for alkalinity}

\item{total_dissolved_solids}{name for TDS to scale color of plot.
Default is NULL}

\item{group_var}{parameter to separate plots by. Default is sample location}

\item{facet_var}{parameter to facet the plot by. Default is sample date}

\item{lines}{TRUE/FALSE plots lines}

\item{cex}{multiplier to scale plot height}
}
\description{
plots a stiff diagram
}
\details{
tidy dataframe of water quality data in mg/L is converted to
meq/L and then transformed to coordinates for Stiff diagmram.
}
\examples{
data(gw_data)
gw_data \%>\%
filter(location_id == "MW-1", sample_date < lubridate::ymd("2008-01-01")) \%>\%
stiff_plot()

gw_data \%>\%
filter(location_id \%in\% c("MW-1", "MW-2")) \%>\%
stiff_plot(., total_dissolved_solids = "Total Dissolved Solids")

}
\keyword{Diagram}
\keyword{Stiff}
\keyword{geochemical}
\keyword{plots}
