% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochemistry_plots.R
\name{stiff_plot}
\alias{stiff_plot}
\title{Function to plot a Stiff Diagram}
\usage{
stiff_plot(df, lines = FALSE, TDS = FALSE, cex = 1)
}
\arguments{
\item{df}{data frame of groundwater data transformed using}

\item{lines}{TRUE/FALSE plots lines}

\item{TDS}{TRUE/FALSE fills in the color of the stiff diagram by TDS}

\item{cex}{multiplier to scale plot height
\code{\link{transform_stiff_data}}}
}
\description{
Function to plot a Stiff Diagram
}

