% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stiff_plot}
\alias{stiff_plot}
\title{Function to plot a Stiff Diagram}
\usage{
stiff_plot(df, TDS = FALSE)
}
\arguments{
\item{df}{data frame of groundwater data transformed using}

\item{TDS}{TRUE/FALSE fills in the color of the stiff diagram by TDS
\code{\link{transform_stiff_data}}}

\item{multiple}{TRUE/FALSE plots multiple locations with facet_wrap}
}
\description{
Function to plot a Stiff Diagram
}

