% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochemistry_plots.R
\name{stiff_time_html}
\alias{stiff_time_html}
\title{function to create an animated Stiff Diagram and save to html}
\usage{
stiff_time_html(df, TDS = FALSE)
}
\arguments{
\item{df}{data frame of groundwater data transformed using}

\item{TDS}{TRUE/FALSE fills polygon with color gradient of TDS
\code{\link{transform_stiff_data}}}
}
\description{
function to create an animated Stiff Diagram and save to html
}
