% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stiff_time_plot}
\alias{stiff_time_plot}
\title{function to create an animated Stiff Diagram}
\usage{
stiff_time_plot(df, TDS = FALSE)
}
\arguments{
\item{df}{data frame of groundwater data transformed using
\code{\link{transform_stiff_data}}}
}
\description{
function to create an animated Stiff Diagram
}

