% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_extract.R
\name{store_inventory}
\alias{store_inventory}
\title{Store an inventory object as a .rda file}
\usage{
store_inventory(inventory_object, filepath)
}
\arguments{
\item{inventory_object}{An inventory object, a data.frame from the `take_inventory` function.}

\item{filepath}{Location to save the inventory}
}
\value{
saved .rda file
}
\description{
Store an inventory object as a .rda file
}
\examples{
two_pack_inventory <- take_inventory(packages = c("ggplot2", "dplyr"))
\dontrun{store_inventory(two_pack_inventory, filepath = "~/Path/To/Folder/filename.rda")}
}
