% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_person_regex.R
\docType{data}
\name{test_names}
\alias{test_names}
\title{Set of names for testing efficacy of non_person_regex}
\format{
A data frame with 232 rows and 2 variables:
\describe{
  \item{name}{the name}
  \item{type}{the type of name, non-person or person}
}
}
\source{
Created by JGG
}
\usage{
test_names
}
\description{
A dataset containing a mix of non-person names (e.g., "Downtown Books," "Supreme Vacuums, LLC") and individuals' names that can be used to test whether the effecicacy of the regex pattern \code{non_person_regex}.
}
\keyword{datasets}
