% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{to_censored}
\alias{to_censored}
\title{Function to convert character column of less than, or greater than
symbols to logical vectors and return a data frame}
\usage{
to_censored(df)
}
\arguments{
\item{df}{data frame of groundwater data}
}
\description{
Function to convert character column of less than, or greater than
symbols to logical vectors and return a data frame
}
