% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camel_snake.R
\name{to_snake}
\alias{to_snake}
\title{to_snake}
\usage{
to_snake(x, nums_to_snake = T)
}
\arguments{
\item{x}{A vector (character or coercible to character).}

\item{nums_to_snake}{Should numbers be snaked as well? ('z1' --> 'z_1')}
}
\value{
A vector.
}
\description{
Converts a vector of strings (usually column names) from camelCase to snake_case.
}
\details{
\code{to_snake()} will return the input with a warning unless there is at least one
capital letter not in the first position somewhere in the set of input strings.

E.g., \code{to_snake(c('aaa', 'bbb', 'Ccc'))} will return \code{x} and a warning.

No dependencies.
}
\examples{
camels <- c('dromedaryCamel', 'wildBactrianCamel', 'BactrianLookingUp',
'aCaravanOf10Camels', 'aLargerCaravanOf11')
to_snake(camels)

}
