% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochemistry_plots.R
\name{transform_piper_data}
\alias{transform_piper_data}
\title{Function to transform data for piper plot}
\usage{
transform_piper_data(df, Mg = "Magnesium, dissolved",
  Ca = "Calcium, dissolved", Na = "Sodium, dissolved",
  K = "Potassium, dissolved", Cl = "Chloride, total",
  SO4 = "Sulfate, total", CO3 = NULL, Alk = "Alkalinity, total (lab)",
  TDS = NULL, name = "location_id", date = "sample_date", units = NULL)
}
\arguments{
\item{df}{data frame of groundwater data}

\item{Mg}{Magnesium}

\item{Ca}{Calcium}

\item{Na}{Sodium}

\item{K}{Potassum}

\item{Cl}{Chloride}

\item{SO4}{Sulfate}

\item{TDS}{Total Dissolved Solids}

\item{name}{column of well names}

\item{date}{column of dates}

\item{units}{right now can handle mg/L and meq/L}

\item{HCO3}{Bicarbonate, or Alkalinity}
}
\description{
Function to transform data for piper plot
}
\keyword{diagram}
\keyword{piper}
