% Generated by roxygen2 (4.0.2): do not edit by hand
\name{transform_stiff_data}
\alias{transform_stiff_data}
\title{Function to transform data from \code{\link{get_major_ions}}
into x, y coordinates and in the
correct path for geom_polygon of ggplot}
\usage{
transform_stiff_data(df, Mg = "Magnesium, dissolved",
  Ca = "Calcium, dissolved", Na = "Sodium, dissolved",
  K = "Potassium, dissolved", Cl = "Chloride, total",
  SO4 = "Sulfate, total", HCO3 = "Alkalinity, total (lab)",
  name = "location_id", date = "sample_date", TDS = NULL, units = NULL)
}
\arguments{
\item{df}{data frame of groundwater data from \code{\link{get_major_ions}}}
}
\description{
Function to transform data from \code{\link{get_major_ions}}
into x, y coordinates and in the
correct path for geom_polygon of ggplot
}

