% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochemistry_plots.R
\name{transform_stiff_data}
\alias{transform_stiff_data}
\title{Function to transform data from \code{\link{get_major_ions}} 
into x, y coordinates and in the 
correct path for geom_polygon of ggplot}
\usage{
transform_stiff_data(df, Mg = "Magnesium, dissolved",
  Ca = "Calcium, dissolved", Na = "Sodium, dissolved",
  K = "Potassium, dissolved", Cl = "Chloride, total",
  SO4 = "Sulfate, total", Alk = "Alkalinity, total (lab)",
  name = "location_id", date = "sample_date", TDS = NULL, units = NULL)
}
\arguments{
\item{df}{data frame of groundwater data from \code{\link{get_major_ions}}}

\item{Mg}{Magnesium}

\item{Ca}{Calcium}

\item{Na}{Sodium,}

\item{K}{Potassium}

\item{Cl}{Chloride}

\item{SO4}{Sulfate}

\item{name}{location_id}

\item{date}{sample_date}

\item{TDS}{Total Dissolved Solids}

\item{units}{Units of data}

\item{HCO3}{Alkalinity}
}
\description{
Function to transform data from \code{\link{get_major_ions}} 
into x, y coordinates and in the 
correct path for geom_polygon of ggplot
}
