% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiff_plot.R
\name{transform_stiff_data}
\alias{transform_stiff_data}
\title{Function to transform data to Stiff diagram coordinates}
\usage{
transform_stiff_data(df, Mg = "Magnesium, dissolved",
  Ca = "Calcium, dissolved", Na = "Sodium, dissolved",
  K = "Potassium, dissolved", Cl = "Chloride, total",
  SO4 = "Sulfate, total", Alk = "Alkalinity, total (lab)",
  name = "LOCATION_ID", date = "SAMPLE_DATE", TDS = NULL, units = NULL)
}
\arguments{
\item{df}{data frame of groundwater data from \code{\link{get_major_ions}}}

\item{Mg}{Magnesium}

\item{Ca}{Calcium}

\item{Na}{Sodium,}

\item{K}{Potassium}

\item{Cl}{Chloride}

\item{SO4}{Sulfate}

\item{name}{LOCATION_ID}

\item{date}{SAMPLE_DATE}

\item{TDS}{Total Dissolved Solids}

\item{units}{Units of data}

\item{HCO3}{Alkalinity}
}
\description{
Function to transform data to Stiff diagram coordinates
}
