\name{transform_stiff_data}
\alias{transform_stiff_data}
\title{Function to transform data from \code{\link{get_plot_data}} into x, y coordinates and in the
correct path for geom_polygon of ggplot}
\usage{
transform_stiff_data(df)
}
\arguments{
  \item{df}{data frame of groundwater data from
  \code{\link{get_plot_data}}}
}
\description{
Function to transform data from \code{\link{get_plot_data}}
into x, y coordinates and in the correct path for
geom_polygon of ggplot
}

