% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_plots.R
\name{ts_plot}
\alias{ts_plot}
\title{This function plots multiple groundwater data time series by location, 
or constituent.}
\usage{
ts_plot(df, facet_by = "location_id", ...)
}
\arguments{
\item{df}{groundwater data}

\item{facet_by}{parameter to group plots by}

\item{back_date}{dates for background date range}

\item{comp_date}{dates for compliance date range}

\item{limit1}{horizontal line 1}

\item{limit2}{horizontal line 2}

\item{short_name}{If TRUE, the analyte name will be abbreviated}

\item{pnt}{size of points on time series plots}

\item{ncol}{number of columns}
}
\description{
This function plots multiple groundwater data time series by location, 
or constituent.
}
