% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_plots.r
\name{ts_plot}
\alias{ts_plot}
\title{This function plots multiple groundwater data time series by location, 
or constituent.}
\usage{
ts_plot(df, x = "sample_date", y = "analysis_result",
  facet_var = "location_id", group_var = "param_name", trend = NULL,
  back_date = NULL, comp_date = NULL, limit1 = NULL, limit2 = NULL,
  short_name = FALSE, pnt = 3, ncol = NULL, ...)
}
\arguments{
\item{df}{df groundwater data in tidy format}

\item{x}{x column for x variable}

\item{y}{y column for y variable}

\item{facet_var}{column to facet wrap plots by, default is by location}

\item{group_var}{column to group plots by, default is by constituent}

\item{trend}{trend add trend line to time series plot}

\item{back_date}{dates for background date range}

\item{comp_date}{dates for compliance date range}

\item{limit1}{horizontal line 1}

\item{limit2}{horizontal line 2}

\item{short_name}{If TRUE, the constituent name will be abbreviated}

\item{pnt}{size of points on time series plots}

\item{ncol}{number of columns}

\item{...}{parameters passed to get_theilsen}
}
\description{
This function plots multiple groundwater data time series by location, 
or constituent.
}
