% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{tukey_outlier}
\alias{tukey_outlier}
\title{Tukey's test for outliers}
\usage{
tukey_outlier(df, x = "analysis_result", k = 3,
  group_by_location = FALSE)
}
\arguments{
\item{df}{data.frame of groundwater data}

\item{x}{column of analysis results}

\item{k}{multiplier for IQR k = 1.5 indicates "outlier", k = 3 indicates "far out"}

\item{group_by_location}{TRUE/FALSE}
}
\description{
Tukey's test for outliers
}
